% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{AddChartRequest}
\alias{AddChartRequest}
\title{AddChartRequest}
\usage{
AddChartRequest(chartSpec, embeddedObjectPosition, chartId = NULL)
}
\arguments{
\item{chartSpec}{a \code{\link[=ChartSpec]{ChartSpec()}} object containing specification for a chart}

\item{embeddedObjectPosition}{an \code{\link[=EmbeddedObjectPosition]{EmbeddedObjectPosition()}} object containing
specification for chart position}

\item{chartId}{integer. optional ID for the chart}
}
\value{
gsheetRequest
}
\description{
Constructs request for creation of googlesheets chart
}
