% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{BasicChartAxis}
\alias{BasicChartAxis}
\alias{is.BasicChartAxis}
\title{BasicChartAxis}
\usage{
BasicChartAxis(title, position = c("BOTTOM_AXIS", "LEFT_AXIS", "RIGHT_AXIS"))

is.BasicChartAxis(x)
}
\arguments{
\item{title}{Title for the axis}

\item{position}{Position of the axis}

\item{x}{any R object}
}
\value{
object of class BasicChartAxis
}
\description{
Specification for googlesheets Basic Chart
}
