% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{BasicChartSeries}
\alias{BasicChartSeries}
\alias{is.BasicChartSeries}
\title{BasicChartSeries}
\usage{
BasicChartSeries(
  gridRange,
  targetAxis = c("BOTTOM_AXIS", "LEFT_AXIS", "RIGHT_AXIS"),
  type = c("LINE", "AREA", "COLUMN"),
  lineStyle = NULL
)

is.BasicChartSeries(x)
}
\arguments{
\item{gridRange}{specification for the cells. Only accepts one-row or
one-column cell ranges}

\item{targetAxis}{specifies on which axis the values will be visualised}

\item{type}{specifies the type of the series (relevant only if chart type is COMBO)}

\item{lineStyle}{specifies style of the line (for LINE type series)}

\item{x}{any R object}
}
\description{
object of class BasicChartDomain. Specifies the data
that will be visualized in the chart
}
