% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{BasicChartSpec}
\alias{BasicChartSpec}
\alias{is.BasicChartSpec}
\title{BasicChartSpec}
\usage{
BasicChartSpec(
  chartType = c("BAR", "LINE", "AREA", "COLUMN", "SCATTER", "COMBO", "STEPPED_AREA"),
  legendPosition = c("BOTTOM_LEGEND", "LEFT_LEGEND", "RIGHT_LEGEND", "TOP_LEGEND",
    "NO_LEGEND"),
  axis,
  domains,
  series,
  headerCount = 0,
  threeDimensional = FALSE
)

is.BasicChartSpec(x)
}
\arguments{
\item{chartType}{type of the chart}

\item{legendPosition}{position of the legend}

\item{axis}{list of 'BasicChartAxis' objects}

\item{domains}{list containing one 'BasicChartDomain' object}

\item{series}{list of 'BasicChartSeries' objects}

\item{headerCount}{The number of rows or columns in the data that are "headers".}

\item{threeDimensional}{boolean indicating if chart need to be made 3D. Only
for charts of types: \code{"BAR"} or \code{"COLUMN"}}

\item{x}{any R object}
}
\description{
Specification for googlesheets Basic Chart object
}
