% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{BasicFilter}
\alias{BasicFilter}
\alias{is.BasicFilter}
\title{BasicFilter}
\usage{
BasicFilter(range, sortSpecs = NULL, filterSpecs = NULL)

is.BasicFilter(x)
}
\arguments{
\item{range}{object of class \link{GridRange}. The range the filter covers.}

\item{sortSpecs}{object of class \link{SortSpec} or list of such objects.
The sort order per column. Later specifications are used when values are
equal in the earlier specifications.}

\item{filterSpecs}{object of class \link{FilterSpec} or list of such objects.
The filter criteria per column.}

\item{x}{any R object}
}
\description{
The default filter associated with a sheet.
}
