% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicCharts.R
\name{BasicSeriesDataPointStyleOverride}
\alias{BasicSeriesDataPointStyleOverride}
\alias{is.BasicSeriesDataPointStyleOverride}
\title{Style override for BasicChartSeries}
\usage{
BasicSeriesDataPointStyleOverride(
  index,
  colorStyle = NULL,
  pointStyle = NULL,
  ...
)

is.BasicSeriesDataPointStyleOverride(x)
}
\arguments{
\item{index}{zero-based index for the series data point}

\item{colorStyle}{object of \link{ColorStyle} class}

\item{pointStyle}{object of \link{PointStyle} class, valid if series (or whole chart)
type is on of: \code{"AREA"}, \code{"LINE"} or \code{"SCATTER"}}

\item{...}{for deprecated GoogleSheets API fields}

\item{x}{any R object}
}
\description{
Style override settings for a single series data point.
}
