% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{BigQueryDataSourceSpec}
\alias{BigQueryDataSourceSpec}
\alias{is.BigQueryDataSourceSpec}
\title{BigQueryDataSourceSpec}
\usage{
BigQueryDataSourceSpec(projectId, querySpec = NULL, tableSpec = NULL)

is.BigQueryDataSourceSpec(x)
}
\arguments{
\item{projectId}{The ID of a BigQuery enabled GCP project with a billing
account attached. For any queries executed against the data source, the project is charged.}

\item{querySpec}{object of class \link{BigQueryQuerySpec}}

\item{tableSpec}{object of class \link{BigQueryTableSpec}}

\item{x}{any R object}
}
\description{
The specification of a BigQuery data source that's connected to a sheet.
Exactly one of \code{querySpec} or \code{tableSpec} can be specified.
}
