% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{BigQueryQuerySpec}
\alias{BigQueryQuerySpec}
\alias{is.BigQueryQuerySpec}
\title{BigQueryQuerySpec}
\usage{
BigQueryQuerySpec(rawQuery)

is.BigQueryQuerySpec(x)
}
\arguments{
\item{rawQuery}{The raw query string.}

\item{x}{any R object}
}
\description{
Specifies a custom BigQuery query.
}
