% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{BigQueryTableSpec}
\alias{BigQueryTableSpec}
\alias{is.BigQueryTableSpec}
\title{BigQueryTableSpec}
\usage{
BigQueryTableSpec(tableId, datasetId, tableProjectId = NULL)

is.BigQueryTableSpec(x)
}
\arguments{
\item{tableId}{The BigQuery table id.}

\item{datasetId}{The BigQuery dataset id.}

\item{tableProjectId}{The ID of a BigQuery project the table belongs to. If not specified, the projectId is assumed.}

\item{x}{any R object}
}
\description{
Specifies a BigQuery table definition. Only native tables is allowed.
}
