% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditions.R
\name{BooleanCondition}
\alias{BooleanCondition}
\alias{is.BooleanCondition}
\title{Boolean Condition}
\usage{
BooleanCondition(
  type = c("NUMBER_GREATER", "NUMBER_GREATER_THAN_EQ", "NUMBER_LESS",
    "NUMBER_LESS_THAN_EQ", "NUMBER_EQ", "NUMBER_NOT_EQ", "NUMBER_BETWEEN",
    "NUMBER_BETWEEN", "NUMBER_NOT_BETWEEN", "TEXT_CONTAINS", "TEXT_NOT_CONTAINS",
    "TEXT_STARTS_WITH", "TEXT_ENDS_WITH", "TEXT_EQ", "TEXT_IS_EMAIL", "DATE_EQ",
    "DATE_BEFORE", "DATE_AFTER", "DATE_ON_OR_BEFORE", "DATE_ON_OR_AFTER", "DATE_BETWEEN",
    "DATE_NOT_BETWEEN", "DATE_IS_VALID", "ONE_OF_RANGE", "ONE_OF_LIST", "BLANK",
    "NOT_BLANK", "CUSTOM_FORMULA", "BOOLEAN", "TEXT_NOT_EQ", 
     "DATE_NOT_EQ"),
  values = NULL
)

is.BooleanCondition(x)
}
\arguments{
\item{type}{Condition type. Description of each condition type meaning can be
found in \emph{details}}

\item{values}{object of class \link{ConditionValue} or list of such objects. Number
of supported values depends on type of the condition: some supporting zero,
other one or two an \code{"ONE_OF_LIST"} - arbitrary number. Information can be
found in \emph{details}}

\item{x}{any R object}
}
\description{
Object representing condition that can evaluate to true or false.
These conditions are useb by conditional formatting, data validation and the
criteria in filters
}
\section{Numeric-specific conditions:}{
\itemize{
\item \strong{NUMBER_GREATER}, \strong{NUMBER_GREATER_THAN_EQ}
\itemize{
\item cell's value must be greater (or equal) than condition's value.
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue}
}
\item \strong{NUMBER_LESS}, \strong{NUMBER_LESS_THAN_EQ}
\itemize{
\item cell's value must be less (or equal) than condition's value.
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue}
}
\item \strong{NUMBER_EQ}, \strong{NUMBER_NOT_EQ}
\itemize{
\item cell's value must be equal (or not equal) to the condition's value
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue}. 1 or more for filters on data source objects.
}
\item \strong{NUMBER_BETWEEN}, \strong{NUMBER_NOT_BETWEEN}
\itemize{
\item cell's value must be (or not be) between two condition values.
\item supported in: data validation, conditional formatting and filters
\item 2 \code{ConditionValue}s
}
}
}

\section{Text-specific conditions:}{
\itemize{
\item \strong{TEXT_CONTAINS}, \strong{TEXT_NOT_CONTAINS}
\itemize{
\item cells' value must contain/not contain the condition's value.
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue}
}
\item \strong{TEXT_STARTS_WITH}, \strong{TEXT_ENDS_WITH}
\itemize{
\item cell's value must start/end with the condition's value.
\item supported in: conditional formatiing and filters
\item 1 \code{ConditionValue}
}
\item \strong{TEXT_EQ}
\itemize{
\item cell's value must be exactly the condition's value
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue}. 1 or more for filters on data source objects.
}
\item \strong{TEXT_IS_EMAIL}
\itemize{
\item cell's value must be a valid email address
\item supported in data validation
\item NO \code{ConditionValue}
}
\item \strong{TEXT_IS_URL}
\itemize{
\item cell's value must be a valid URL
\item supported in data validation
\item NO \code{ConditionValue}
}
\item \strong{TEXT_NOT_EQ}
\itemize{
\item cell's value must be exactly not the condition's value.
\item supported by filters on data source objects
\item 1 or more \code{ConditionValue}s
}
}
}

\section{Dates-specific conditions:}{
\itemize{
\item \strong{DATE_EQ}:
\itemize{
\item cell's value must be the same date as condition's value
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue}. 1 or more for filters on data source objects.
}
\item \strong{DATE_BEFORE}, \strong{DATE_AFTER}
\itemize{
\item cell's value must be before (or after) the date of the condition's
value
\item supported in: data validation, conditional formatting and filters
\item 1 \code{ConditionValue} (can be a \emph{relative date})
}
\item \strong{DATE_ON_OR_BEFORE}, \strong{DATE_ON_OR_AFTER}
\itemize{
\item cell's value must be on or before (or after) the date of the condition's
value
\item supported in data validation
\item 1 \code{ConditionValue} (can be a \emph{relative date})
}
\item \strong{DATE_BETWEEN}, \strong{DATE_NOT_BETWEEN}
\itemize{
\item cell's value must be between (or outside of) two condition values.
\item supported in data validation
\item 2 \code{ConditionValue}s
}
\item \strong{DATE_IS_VALID}
\itemize{
\item cell's value must be a date
\item supported in data validation
\item NO \code{ConditionValue}
}
}
}

\section{Non-specific conditions}{
\itemize{
\item \strong{ONE_OF_RANGE}
\itemize{
\item cell's value must be listed in the grid in the condition value's range.
\item supported in data validation
\item 1 \code{ConditionValue} that is a valid range in \emph{A1 notation} (see: \code{\link[=get_A1_not]{get_A1_not()}})
}
\item \strong{ONE_OF_LIST}
\itemize{
\item cell's value must be in the list of condition values
\item supported in data validation
\item Any number of \code{ConditionValue}s
}
\item \strong{BLANK}, \strong{NOT_BLANK}
\itemize{
\item cell's value must be empty (or not empty)
\item supported in conditional formatting and filters
\item NO \code{ConditionValue}
}
\item \strong{CUSTOM_FORMULA}
\itemize{
\item formula in condition value must evaluate to \code{TRUE}
\item supported in data validation, conditional formatting and filters only
in \code{GRID} sheets
\item 1 \code{ConditionValue} with formula
}
\item \strong{BOOLEAN}
\itemize{
\item the cell's value must be TRUE/FALSE or in list of condition values
\item supported in data validation
\item with NO \code{ConditionValue}s: value must be TRUE/FALSE and TRUE renders
as checked, FALSE as unchecked
\item 1 \code{ConditionValue}: cell will render as checked when it contains that value
and unchecked when blank
\item 2 \code{ConditionValue}s: cell will render as checked when it contains first
value and unchecked when contains second
}
}
}

