% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{Border}
\alias{Border}
\alias{Borders}
\alias{is.Border}
\alias{is.Borders}
\title{Border specification}
\usage{
Border(
  style = c("SOLID", "SOLID_MEDIUM", "SOLID_THICK", "DOTTED", "DASHED", "NONE", "DOUBLE"),
  colorStyle = NULL
)

Borders(top = NULL, bottom = NULL, left = NULL, right = NULL)

is.Border(x)

is.Borders(x)
}
\arguments{
\item{style}{Type of the border.}

\item{colorStyle}{object of class \link{ColorStyle} describing color of the border}

\item{top, bottom, left, right}{Objects of class \link{Border} declaring style
of specified border in a cell}

\item{x}{any R object}

\item{...}{additional deprecated fields returned from GoogleSheet API}
}
\description{
Border specification
}
\details{
\strong{Border styles}:
\itemize{
\item \code{SOLID}: The border is a thin solid line.
\item \code{SOLID_MEDIUM}: The border is a medium solid line.
\item \code{SOLID_THICK}: The border is a thick solid line.
\item \code{DOTTED}: The border is dotted.
\item \code{DASHED}: The border is dashed.
\item \code{NONE}: No border. Used only when updating a border in order to erase it.
\item \code{DOUBLE}: The border is two solid lines.
}
}
