% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{Borders}
\alias{Borders}
\title{Cell border specification}
\usage{
Borders(
  top_style = c("SOLID", "DOTTED", "DASHED", "SOLID_MEDIUM", "SOLID_THICK", "NONE",
    "DOUBLE"),
  bottom_style = c("SOLID", "DOTTED", "DASHED", "SOLID_MEDIUM", "SOLID_THICK", "NONE",
    "DOUBLE"),
  left_style = c("SOLID", "DOTTED", "DASHED", "SOLID_MEDIUM", "SOLID_THICK", "NONE",
    "DOUBLE"),
  right_style = c("SOLID", "DOTTED", "DASHED", "SOLID_MEDIUM", "SOLID_THICK", "NONE",
    "DOUBLE"),
  top_colorStyle = NULL,
  bottom_colorStyle = NULL,
  left_colorStyle = NULL,
  right_colorStyle = NULL
)
}
\arguments{
\item{top_style, bottom_style, left_style, right_style}{Type of the top, bottom,
left and right borders.}

\item{top_colorStyle, bottom_colorStyle, left_colorStyle, right_colorStyle}{Object
of class \code{\link[=ColorStyle]{ColorStyle()}} describing color for top, bottom, left and right
border}
}
\description{
Cell border specification
}
\details{
\strong{Border styles}:
\itemize{
\item \code{DOTTED}: The border is dotted.
\item \code{DASHED}: The border is dashed.
\item \code{SOLID}: The border is a thin solid line.
\item \code{SOLID_MEDIUM}: The border is a medium solid line.
\item \code{SOLID_THICK}: The border is a thick solid line.
\item \code{NONE}: No border. Used only when updating a border in order to erase it.
\item \code{DOUBLE}: The border is two solid lines.
}
}
