% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dgs4_cli.R
\name{bulletize_fields}
\alias{bulletize_fields}
\title{cli utility for creating bullet point fields list}
\usage{
bulletize_fields(fields, fields_names = NULL)
}
\description{
cli utility for creating bullet point fields list
}
