% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{CellData}
\alias{CellData}
\alias{is.CellData}
\alias{gen_CellData}
\title{Cell Data}
\usage{
CellData(
  userEnteredValue = NULL,
  userEnteredFormat = NULL,
  note = NULL,
  textFormatRuns = NULL,
  dataValidation = NULL,
  pivotTable = NULL,
  effectiveValue = NULL,
  formattedValue = NULL,
  effectiveFormat = NULL,
  hyperlink = NULL,
  ...
)

is.CellData(x)

gen_CellData(obj)
}
\arguments{
\item{userEnteredValue}{object of class \link{ExtendedValue}. Value that user
entered in the cell.}

\item{userEnteredFormat}{object of class \link{CellFormat}. Format that user
provided for the cell. During write, new format will be merged with
existing format.}

\item{note}{Note entered on the cell.}

\item{textFormatRuns}{List of objects of class \link{TextFormatRun}. Sequence
of formats if parts of the value needs to have different formatting}

\item{dataValidation}{object of class \link{DataValidationRule}. Specification
for data validation for a cell. During write, new data validation rule will
overwrite prior rule}

\item{pivotTable}{object of class \link{PivotTable}. Pivot table anchored at this
cell. Size of pivot table is computed dynamically, only the top-left cell
contains definition of pivot table: all other cells will contain calculated
values of the results in their \code{effectiveValue} fields}

\item{effectiveValue}{\strong{ReadOnly} Object of class \link{ExtendedValue}.
For cells with formulas: calculated value. For cells with literals,
the same as \code{userEnteredValue}.}

\item{formattedValue}{\strong{ReadOnly} Formatted value of the cell: Value as
it's shown to the user. \strong{ReadOnly}}

\item{effectiveFormat}{\strong{ReadOnly} Object of class \link{CellFormat}. Effective
format of the cell: results of conditional formatting, computed number
format if cell contains formula. If it would be the same as \code{userEnteredFormat},
it won't be received}

\item{hyperlink}{\strong{ReadOnly} Hyperlink this cell points to, if any. For
multiple hyperlinks: it would be empty. To set hyperlink, you can use
\code{"=HYPERLINK"} formula for \link{ExtendedValue} provided to \code{userEnteredValue}
or link with \link{textFormat} provided to \code{userEnteredFormat} (for one link)
or in \code{textFormatRuns} for multiple links or link spanning part of the value}

\item{...}{Other}

\item{x}{any R object}

\item{obj}{list produced by \code{deepgs_listinize()}}
}
\description{
Specification of data contained in singular sheet cell
}
