% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{CellFormat}
\alias{CellFormat}
\alias{gen_CellFormat}
\title{Cell Format}
\usage{
CellFormat(
  numberFormat = NULL,
  backgroundColorStyle = NULL,
  borders = NULL,
  padding = NULL,
  horizontalAlignment = NULL,
  verticalAlignment = NULL,
  wrapStrategy = NULL,
  textDirection = NULL,
  textFormat = NULL,
  hyperlinkDisplayType = NULL,
  textRotation = NULL,
  ...
)

gen_CellFormat(obj)
}
\arguments{
\item{numberFormat}{object of \link{NumberFormat} class. Describes how the
numbers are to be formatted within the cells}

\item{borders}{object of \link{Borders} class, describing the borders of the cell}

\item{padding}{object of \link{Padding} class, describing the padding around the cell}

\item{horizontalAlignment, verticalAlignment}{horizontal and vertical alignments
of values in the cells}

\item{wrapStrategy}{strategy for wrapping values in the cell}

\item{textDirection}{direction of the text in the cell}

\item{textFormat}{object of \link{TextFormat} class, describing format of the text
in cells}

\item{hyperlinkDisplayType}{how hyperlinks should be displayed in cells}

\item{textRotation}{rotation applied to text in the cell. Either \code{v} for vertical
or integer between \code{-90} and \code{90} for angle of rotation}

\item{...}{additional deprecated fields returned from GoogleSheet API}

\item{obj}{list produced by \code{deepgs_listinize()}}

\item{backgrundColorStyle}{object of \link{ColorStyle} class, describing the color
of cell background}
}
\description{
Specification for format of the cells
}
\details{
Valid values for enumerative fields:
\itemize{
\item \strong{verticalAlignment}:
\itemize{
\item \code{TOP},
\item \code{MIDDLE},
\item \code{BOTTOM}
}
\item \strong{horizontalAlignment}
\itemize{
\item \code{LEFT}
\item \code{CENTER}
\item \code{RIGHT}
}
\item \strong{wrapStrategy}
\itemize{
\item \code{OVERFLOW_CELL}: lines that are longer than the cell width will be written
in the next cell over, unless it isn't emppty - then the same behaviour as
in \code{CLIP} is used
\item \code{LEGACY_WRAP}: old GoogleSheets strategy, not supported on all platforms.
Words wider than cell width are clipped, not broken. Its usage is discouraged.
\item \code{CLIP}: Lines longer than cell width will be clipped.
\item \code{WRAP}: Words longer than cell width are wrapped into new line.
}
\item \strong{textDirection}:
\itemize{
\item \code{LEFT_TO_RIGHT}
\item \code{RIGHT_TO_LEFT}
}
\item \strong{hyperlingDisplayType}:
\itemize{
\item \code{LINKED}: hyperlink is explicitly rendered
\item \code{PLAIN_TEXT}: hyperlink is not rendered
}
}
}
