% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{ChartData}
\alias{ChartData}
\title{ChartData}
\usage{
ChartData(
  sourceRange = NULL,
  columnReference = NULL,
  aggregateType = NULL,
  groupRule = NULL
)
}
\arguments{
\item{sourceRange}{object of class \link{GridRange} or list of such objects.
All of them need to have one of the dimension of length one (one column or one row).
need to be provided for non-data source charts.}

\item{columnReference}{Display name of the data source column that the data
reads from. Need to be provided for data source charts.}

\item{aggregateType}{aggregation type for the \emph{series} of a data source chart.
Only for data-source charts}

\item{groupRule}{object of class \code{ChartGroupRule}. Only for data source charts.}
}
\description{
Specification for data to be used for creation of \emph{domain},
\emph{series} or other specific types
}
