% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{ChartGroupRule}
\alias{ChartGroupRule}
\alias{is.ChartGroupRule}
\title{ChartGroupRule}
\usage{
ChartGroupRule(dateTimeRule = NULL, histogramRule = NULL)

is.ChartGroupRule(x)
}
\arguments{
\item{dateTimeRule}{Allows you to organize the date-time values in a source
data column into buckets based on selected parts of their date or time values.}

\item{histogramRule}{object of class \link{ChartHistogramRule}. Allows
organizing numeric values in a source data column into buckets of constant size.}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{ChartGroupRule}
}
\description{
An optional setting on the \link{ChartData} of the \strong{domain} of a
\link{DataSource}-based chart that defines buckets for the values in the domain
rather than breaking out each individual value.

Either \code{dateTimeRule} or \code{histogramRule} needs to be specified

For example, when plotting a data source chart, you can specify a histogram
rule on the domain (it should only contain numeric values), grouping its
values into buckets. Any values of a chart series that fall into the same
bucket are aggregated based on the \code{aggregateType}
}
\section{dateTimeRule options}{

\strong{SECOND} Group dates by second, from \code{0} to \code{59}.
\strong{MINUTE} Group dates by minute, from \code{0} to \code{59}.
\strong{HOUR} Group dates by hour using a 24-hour system, from \code{0} to \code{23}.
\strong{HOUR_MINUTE} Group dates by hour and minute using a 24-hour system,
for example \code{19:45}.
\strong{HOUR_MINUTE_AMPM} Group dates by hour and minute using a 12-hour system,
for example \verb{7:45 PM}. The AM/PM designation is translated based on the
spreadsheet locale.
\strong{DAY_OF_WEEK} Group dates by day of week, for example \code{Sunday}. The days
of the week will be translated based on the spreadsheet locale.
\strong{DAY_OF_YEAR} Group dates by day of year, from \code{1} to \code{366}. Note that dates
after \verb{Feb. 29} fall in different buckets in leap years than in non-leap years.
\strong{DAY_OF_MONTH} Group dates by day of month, from \code{1} to \code{31}.
\strong{DAY_MONTH} Group dates by day and month, for example \code{22-Nov}. The month
is translated based on the spreadsheet locale.
\strong{MONTH} Group dates by month, for example \code{Nov}. The month is translated
based on the spreadsheet locale.
\strong{QUARTER} Group dates by quarter, for example \code{Q1} (which represents \emph{Jan-Mar}).
\strong{YEAR} Group dates by year, for example \code{2008}.
\strong{YEAR_MONTH} Group dates by year and month, for example \code{2008-Nov}. The
month is translated based on the spreadsheet locale.
\strong{YEAR_QUARTER} Group dates by year and quarter, for example \verb{2008 Q4}.
\strong{YEAR_MONTH_DAY} Group dates by year, month, and day, for example \code{2008-11-22}.
}

\seealso{
Other Chart objects constructors: 
\code{\link{ChartAxisViewWindowOptions}()},
\code{\link{ChartData}()},
\code{\link{ChartHistogramRule}()},
\code{\link{ChartSpec}()},
\code{\link{DataLabel}()},
\code{\link{DataSourceChartProperties}()},
\code{\link{EmbeddedChart}()}
}
\concept{Chart objects constructors}
