% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{ChartHistogramRule}
\alias{ChartHistogramRule}
\alias{is.ChartHistogramRule}
\title{ChartHistogramRule}
\usage{
ChartHistogramRule(intervalSize, minValue = NULL, maxValue = NULL)

is.ChartHistogramRule(x)
}
\arguments{
\item{intervalSize}{The size of the buckets that are created. Must be positive.}

\item{minValue, maxValue}{The minimum and maximum values at which items are
placed into buckets. Values outside of this range will be gathered in a
single, additional bucket on that side. If omitted, it is determined by the
minimum or maximum item value.}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{ChartHistogramRule}
}
\description{
Allows you to organize numeric values when creating a \link{ChartGroupRule}
for a \link{DataSource} chart
}
\seealso{
Other Chart objects constructors: 
\code{\link{ChartAxisViewWindowOptions}()},
\code{\link{ChartData}()},
\code{\link{ChartGroupRule}()},
\code{\link{ChartSpec}()},
\code{\link{DataLabel}()},
\code{\link{DataSourceChartProperties}()},
\code{\link{EmbeddedChart}()}
}
\concept{Chart objects constructors}
