% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChartRequests.R
\name{ChartRequests}
\alias{ChartRequests}
\alias{AddChartRequest}
\alias{UpdateChartSpecRequest}
\title{Requests to add and update Chart}
\usage{
AddChartRequest(chart)

UpdateChartSpecRequest(chartId, spec)
}
\arguments{
\item{chart}{an \link{EmbeddedChart} object containing specification for a chart}

\item{chartId}{id of the chart to update}

\item{spec}{object of class \link{ChartSpec}, containing new chart specifications}
}
\value{
dgs4Req object
}
\description{
Create \code{dgs4Req} objects that allow for addition or modification of chart
embedded in a spreadsheet. Send created requests with \code{\link[=request_ss_batchUpdate]{request_ss_batchUpdate()}}
}
\section{AddChart}{

Adds a chart to spreadsheet.
}

\section{UpdateChartSpec}{

Constructs request for updating chart specification of existing chart
}

\seealso{
Other dgs4Req constructors: 
\code{\link{DimensionRequests}},
\code{\link{PasteRequests}},
\code{\link{SheetRequests}}
}
\concept{dgs4Req constructors}
