% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{ChartSpec}
\alias{ChartSpec}
\alias{is.ChartSpec}
\title{ChartSpec}
\usage{
ChartSpec(
  chart,
  title = NULL,
  titleTextPosition = NULL,
  titleTextFormat = NULL,
  subtitle = NULL,
  subtitleTextPosition = NULL,
  subtitleTextFormat = NULL,
  fontName = NULL,
  altText = NULL,
  maximized = NULL,
  backgroundColorStyle = NULL,
  dataSourceChartProperties = NULL,
  filterSpecs = NULL,
  sortSpecs = NULL,
  hiddenDimensionStrategy = NULL,
  ...
)

is.ChartSpec(x)
}
\arguments{
\item{chart}{Specification of the chart to render. One of \link{TypeChartSpec}.}

\item{title, subtitle}{Title and subtitle of the chart.}

\item{titleTextPosition, subtitleTextPosition}{Horizntal aligmnent of the title and
subtitle.}

\item{titleTextFormat, subtitleTextFormat}{Objects of \link{TextFormat} class.
Strikethrough, underline and link aren't supported.}

\item{fontName}{name of the font to use.}

\item{altText}{alternative text to describe a chart.}

\item{maximized}{boolean. If \code{TRUE}, the chart will take maximum amount of
available space with minimal padding.}

\item{backgroundColorStyle}{object of \link{ColorStyle} class, describing the
background color for entire chart.}

\item{dataSourceChartProperties}{object of \link{DataSourceChartPoperties},
contained only in \emph{data source} charts #NOT DONE YET#}

\item{filterSpecs}{object or list of objects of \link{FilterSpec} class, describing
filters applied to the source data. Only in \emph{data source} charts. #NOT DONE YET#}

\item{sortSpecs}{object of \link{SortSpec} class, describing the order to sort
the chart data. Only in \emph{data source} charts. #NOT DONE YET#}

\item{hiddenDimensionStrategy}{describing how the charts will use hidden
rows or columns.}

\item{...}{for support of deprecated GoogleSheets API fields during read}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{ChartSpec}
}
\description{
Specification for googlesheets chart
}
\seealso{
Other Chart objects constructors: 
\code{\link{ChartAxisViewWindowOptions}()},
\code{\link{ChartData}()},
\code{\link{DataLabel}()},
\code{\link{DataSourceChartProperties}()},
\code{\link{EmbeddedChart}()}
}
\concept{Chart objects constructors}
