% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varia.R
\name{ColorStyle}
\alias{ColorStyle}
\alias{is.ColorStyle}
\title{ColorStyle}
\usage{
ColorStyle(
  red,
  green,
  blue,
  alpha = NULL,
  themeColorType = c("TEXT", "BACKGROUND", "ACCENT1", "ACCENT2", "ACCENT3", "ACCENT4",
    "ACCENT5", "ACCENT6", "LINK")
)

is.ColorStyle(x)
}
\arguments{
\item{red, green, blue}{Floats between \code{0} and \code{1}, describing amount of given
tone in the color. If all are given, they will take precedence before \code{themeColorType}}

\item{alpha}{Float between \code{0} and \code{1}. Not universally valid through the
Sheets API}

\item{themeColorType}{One of the colors set in the theme of the spreadsheet.
See \strong{details} for more info.}

\item{x}{Any R object}
}
\value{
ColorStyle object
}
\description{
ColorStyle
}
\details{
ColorStyle can be represented in two ways: manually, by providing
\code{red}, \code{green}, \code{blue} (and, optionally \code{alpha}) values describing
the color, or dynamically, by assigning one of \code{themeColorType}s, which are
set in the \link{SpreadsheetProperties}. Valid styles:
\itemize{
\item \emph{TEXT}: Represents the primary text color
\item \emph{BACKGROUND}: Represents the primary background color
\item \emph{ACCENT1} to \emph{ACCENT6}: Represents accents color
\item \emph{LINK}: Represents the color to use for hyperlinks
}
}
