% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFormatRequests.R
\name{ConditionalFormatRequests}
\alias{ConditionalFormatRequests}
\alias{AddConditionalFormatRule}
\alias{UpdateConditionalFormatRule}
\alias{DeleteConditionalFormatRule}
\alias{AddConditionalFormatRuleRequest}
\alias{UpdateConditionalFormatRuleRequest}
\alias{DeleteConditionalFormatRuleRequest}
\title{Requests to add, modify and delete conditional formatting}
\usage{
AddConditionalFormatRuleRequest(index, rule)

UpdateConditionalFormatRuleRequest(
  index,
  rule = NULL,
  newIndex = NULL,
  sheetId = NULL
)

DeleteConditionalFormatRuleRequest(index, sheetId)
}
\arguments{
\item{index}{zero-based index identifying the rule}

\item{rule}{object of class \link{ConditionalFormatRule} describing the rule
for addition or update}

\item{newIndex}{zero-based index to which the rule should be moved}

\item{sheetId}{ID of the sheet the rule is located in}
}
\value{
deepgs4Req object
}
\description{
Create \code{deepgs4Req} objects that allow for addition, modification
and deletion of conditional formatting from given sheet. Send created
requests with \code{\link[=request_ss_batchUpdate]{request_ss_batchUpdate()}}
}
\section{Add}{

New rule is added at a given index. All subsequent rules indices are incremented.
}

\section{Update}{

Rule at a given index can be either replaced, providing the \code{rule} argument
or moved, providing the \code{newIndex} and \code{sheetId} arguments.
}

\section{Delete}{

Conditional format rule at given \code{index} from specified \code{sheetId} is
deleted, and all subsequent rules' indices are decremented.
}

\concept{deepgs4Req constructors}
