% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conditions.R
\name{ConditionValue}
\alias{ConditionValue}
\alias{is.ConditionValue}
\title{Condition Value}
\usage{
ConditionValue(userEnteredValue = NULL, relativeDate = NULL)

is.ConditionValue(x)
}
\arguments{
\item{userEnteredValue}{Value this condition is based on. Should be used
for most condition types. Formulas are supported and must begin with
\code{=} or \code{+}}

\item{relativeDate}{Date relative to the current date. Can be used only by
condition types: \code{DATE_BEFORE}, \code{DATE_AFTER}, \code{DATE_ON_OR_BEFORE}, \code{DATE_ON_OR_AFTER}.
Can be one of: \code{"PAST_YEAR"}, \code{"PAST_MONTH"}, \code{"PAST_WEEK"}, \code{"YESTERDAY"},
\code{"TODAY"}, \code{"TOMORROW"}}

\item{x}{any R object}
}
\description{
Value of the condition. Used in \code{\link[=BooleanCondition]{BooleanCondition()}}. Only
one of arguments below can be used.
}
