% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetatadataRequest.R
\name{CreateDeveloperMetadataRequest}
\alias{CreateDeveloperMetadataRequest}
\title{Create metadata request}
\usage{
CreateDeveloperMetadataRequest(developerMetadata)
}
\arguments{
\item{developerMetadata}{object of class \link{DeveloperMetadata}}
}
\description{
Create metadata request
}
