% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{DataExecutionStatus}
\alias{DataExecutionStatus}
\alias{is.DataExecutionStatus}
\title{DataExecutionStatus}
\usage{
DataExecutionStatus(
  state = c("NOT_STARTED", "RUNNING", "SUCCEEDED", "FAILED"),
  errorCode = NULL,
  errorMessage = NULL,
  lastRefreshTime = NULL
)

is.DataExecutionStatus(x)
}
\arguments{
\item{state}{The state of the data execution.}

\item{errorCode}{The error code.}

\item{errorMessage}{errorMessage}

\item{lastRefreshTime}{Gets the time the data last successfully refreshed.}

\item{x}{any R object}
}
\description{
Object returned from GoogleSheet API with details of current
execution of data retrieval for \strong{DATA_SOURCE} Sheet
}
\details{
\subsection{Possible states:}{
\itemize{
\item \strong{NOT_STARTED}: The data execution has not started.
\item \strong{RUNNING}: The data execution has started and is running.
\item \strong{SUCCEEDED}: The data execution has completed successfully.
\item \strong{FAILED}: The data execution has completed with errors.
}
}

\subsection{Possible \code{errorCode}s:}{
\itemize{
\item \strong{TIMED_OUT}: The data execution timed out.
\item \strong{TOO_MANY_ROWS}: The data execution returns more rows than the limit.
\item \strong{TOO_MANY_COLUMNS}: The data execution returns more columns than the limit.
\item \strong{TOO_MANY_CELLS}: The data execution returns more cells than the limit.
\item \strong{ENGINE}: Error is received from the backend data execution engine
(e.g. BigQuery). Check errorMessage for details.
\item \strong{PARAMETER_INVALID}: One or some of the provided data source parameters
are invalid.
\item \strong{UNSUPPORTED_DATA_TYPE}: The data execution returns an unsupported data type.
\item \strong{DUPLICATE_COLUMN_NAMES}: The data execution returns duplicate column
names or aliases.
\item \strong{INTERRUPTED}: The data execution is interrupted. Please refresh later.
\item \strong{CONCURRENT_QUERY}: The data execution is currently in progress, can
not be refreshed until it completes.
\item \strong{OTHER}: Other errors.
\item \strong{TOO_MANY_CHARS_PER_CELL}: The data execution returns values that exceed
the maximum characters allowed in a single cell.
\item \strong{DATA_NOT_FOUND}:	The database referenced by the data source is not found.
\item \strong{PERMISSION_DENIED}: The user does not have access to the database
referenced by the data source.
\item \strong{MISSING_COLUMN_ALIAS}: The data execution returns columns with missing
aliases.
\item \strong{OBJECT_NOT_FOUND}: The data source object does not exist.
\item \strong{OBJECT_IN_ERROR_STATE}: The data source object is currently in error
state. To force refresh, set force in RefreshDataSourceRequest .
\item \strong{OBJECT_SPEC_INVALID}: The data source object specification is invalid.
}
}
}
