% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetadata.R
\name{DataFilter}
\alias{DataFilter}
\alias{is.DataFilter}
\title{Data Filter}
\usage{
DataFilter(developerMetadataLookup = NULL, a1Range = NULL, gridRange = NULL)

is.DataFilter(x)
}
\arguments{
\item{developerMetadataLookup}{object of class \link{DeveloperMetadataLookup}.
Selects data associated with the developer metadata matching the criteria.}

\item{a1Range}{Selects data that matches the specified A1 range. Needs
to be a valid \emph{A1 range}}

\item{gridRange}{object of class \link{GridRange}
Selects data that matches the specified grid range.}

\item{x}{any R object}
}
\value{
object of class \code{DataFilter}
}
\description{
Filter that describes what data should be selected or returned
from a request. Only one of following arguments can be specified.
}
