% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{DataSource}
\alias{DataSource}
\alias{is.DataSource}
\title{Data Source}
\usage{
DataSource(
  dataSourceId = NULL,
  spec = NULL,
  calculatedColumns = NULL,
  sheetId = NULL
)

is.DataSource(x)
}
\arguments{
\item{dataSourceId}{The spreadsheet-scoped unique ID that identifies the data source.}

\item{spec}{object of \link{DataSourceSpec} class. Specifies data source connected
with this spreadsheet.}

\item{calculatedColumns}{object of class \link{DataSourceColumn} or list of such.
All calculated columns in the data source.}

\item{sheetId}{The ID of the Sheet connected with the data source. The field
cannot be changed once set.}

\item{x}{any R object}
}
\description{
Information about an external data source in the spreadsheet.
}
