% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{DataSourceChartProperties}
\alias{DataSourceChartProperties}
\alias{is.DataSourceChartProperties}
\title{DataSourceChartProperties}
\usage{
DataSourceChartProperties(dataSourceId, dataExecutionStatus = NULL)

is.DataSourceChartProperties(x)
}
\arguments{
\item{dataSourceId}{ID of the data source that the chart is associated with.}

\item{dataExecutionStatus}{**READ ONLY ** object of class \link{DataExecutionStatus}.}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{DataSourceChartProperties}
}
\description{
Properties of a chart based on data located in \link{DataSource}.
}
\seealso{
Other Chart objects constructors: 
\code{\link{ChartAxisViewWindowOptions}()},
\code{\link{ChartData}()},
\code{\link{ChartGroupRule}()},
\code{\link{ChartHistogramRule}()},
\code{\link{ChartSpec}()},
\code{\link{DataLabel}()},
\code{\link{EmbeddedChart}()}
}
\concept{Chart objects constructors}
