% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{DataSourceColumn}
\alias{DataSourceColumn}
\alias{is.DataSourceColumn}
\title{DataSourceColumn}
\usage{
DataSourceColumn(reference, formula = NULL)

is.DataSourceColumn(x)
}
\arguments{
\item{reference}{The display name of the column. It should be unique within a data source.}

\item{formula}{The formula of the calculated column.}

\item{x}{any R object}
}
\description{
A column in a \link{DataSource}.
}
