% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{DataSourceRefreshSchedule}
\alias{DataSourceRefreshSchedule}
\alias{is.DataSourceRefreshSchedule}
\title{DataSourceRefreshSchedule}
\usage{
DataSourceRefreshSchedule(
  enabled = NULL,
  startTime = NULL,
  daysOfWeek = NULL,
  daysOfMonth = NULL,
  nextRun = NULL,
  ...
)

is.DataSourceRefreshSchedule(x)
}
\arguments{
\item{enabled}{\code{TRUE} if the refresh schedule is enabled, or false otherwise.}

\item{startTime}{object of class \link{TimeOfDay} specifying the startTime of schedule
refresh. If neither \code{daysOfWeek} or \code{daysOfMonth} are specified, refresh will
be made daily}

\item{daysOfWeek}{day of the week when the refresh should be made.}

\item{daysOfMonth}{days of month when the refresh should be made. Integers
between 1 and 28 inclusive are valid.}

\item{nextRun}{\strong{READ ONLY} The time interval of the next run. Object of
class \link{Interval}}

\item{...}{for support of some redundant SheetsAPI fields}

\item{x}{any R object}
}
\description{
Schedule for refreshing the \link{DataSource}.

At most one of \code{daysOfWeek} or \code{daysOfMonth} can be specified.
}
\details{
\subsection{daysOfWeek options:}{

MONDAY, TUESDAY, WEDNESDAY, THURSDAY, FRIDAY, SATURDAY, SUNDAY
}

\subsection{Exact refresh time}{

Data sources in the spreadsheet are refreshed within a time interval. You
can specify the start time by clicking the Scheduled Refresh button in the
Sheets editor, but the interval is fixed at 4 hours. For example, if you
specify a start time of 8am , the refresh will take place between 8am and
12pm every day.
}
}
