% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sheet.R
\name{DataSourceSheetProperties}
\alias{DataSourceSheetProperties}
\alias{is.DataSourceSheetProperties}
\title{DataSourceSheetProperties}
\usage{
DataSourceSheetProperties(
  dataSourceId,
  columns = NULL,
  dataExecutionStatus = NULL
)

is.DataSourceSheetProperties(x)
}
\arguments{
\item{dataSourceId}{ID of the \link{DataSource} the \link{Sheet} is connected to}

\item{columns}{object of class \link{DataSourceColumn} or list of such. The columns
displayed on the sheet, corresponding to the values in RowData}

\item{dataExecutionStatus}{object of class \link{DataExecutionStatus}. The data
execution status.}

\item{x}{any R object}
}
\description{
DataSourceSheetProperties
}
