% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{DataSourceSpec}
\alias{DataSourceSpec}
\alias{is.DataSourceSpec}
\title{DataSourceSpec}
\usage{
DataSourceSpec(bigQuery = NULL, parameters = NULL, ...)

is.DataSourceSpec(x)
}
\arguments{
\item{bigQuery}{object of class \link{BigQueryDataSourceSpec}}

\item{parameters}{object of class \link{DataSourceParameter} or list of such.
The parameters of the data source, used when querying the data source.}

\item{...}{other sources data}

\item{x}{any R object}
}
\description{
This specifies the details of the \link{DataSource}. For example,
for BigQuery, this specifies information about the BigQuery source.
}
