% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{deepgs_auth_configure}
\alias{deepgs_auth_configure}
\alias{deepgs_api_key}
\alias{deepgs_oauth_app}
\title{Edit and view auth configuration}
\usage{
deepgs_auth_configure(app, path, api_key)

deepgs_api_key()

deepgs_oauth_app()
}
\arguments{
\item{app}{OAuth app, in the sense of \code{\link[httr:oauth_app]{httr::oauth_app()}}.}

\item{path}{JSON downloaded from Google Cloud Platform Console, containing a
client id (aka key) and secret, in one of the forms supported for the \code{txt}
argument of \code{\link[jsonlite:fromJSON]{jsonlite::fromJSON()}} (typically, a file path or JSON string).}

\item{api_key}{API key.}
}
\value{
\itemize{
\item \code{deepgs_auth_configure()}: An object of R6 class
\link[gargle:AuthState-class]{gargle::AuthState}, invisibly.
\item \code{deepgs_oauth_app()}: the current user-configured
\code{\link[httr:oauth_app]{httr::oauth_app()}}.
\item \code{deepgs_api_key()}: the current user-configured API key.
}
}
\description{
These functions give more control over and visibility into the auth
configuration than \code{\link[=deepgs_auth]{deepgs_auth()}} does. \code{deepgs_auth_configure()}
lets the user specify their own:
\itemize{
\item OAuth app, which is used when obtaining a user token.
\item API key. If deepgsheets4 is de-authorized via \code{\link[=deepgs_deauth]{deepgs_deauth()}}, all
requests are sent with an API key in lieu of a token.
See the vignette
\href{https://gargle.r-lib.org/articles/get-api-credentials.html}{How to get your own API credentials}
for more.
If the user does not configure these settings, internal defaults
are used.
\code{deepgs_oauth_app()} and \code{deepgs_api_key()} retrieve the
currently configured OAuth app and API key, respectively.
}
}
\examples{
# see and store the current user-configured OAuth app (probably `NULL`)
(original_app <- deepgs_oauth_app())

# see and store the current user-configured API key (probably `NULL`)
(original_api_key <- deepgs_api_key())

if (require(httr)) {
  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "YOUR_CLIENT_ID_GOES_HERE",
    secret = "YOUR_SECRET_GOES_HERE"
  )
  google_key <- "YOUR_API_KEY"
  deepgs_auth_configure(app = google_app, api_key = google_key)

  # confirm the changes
  deepgs_oauth_app()
  deepgs_api_key()

  # bring your own app via JSON downloaded from Google Developers Console
  # this file has the same structure as the JSON from Google
  app_path <- system.file(
    "extdata", "fake-oauth-client-id-and-secret.json",
    package = "deepgsheets4"
  )
  deepgs_auth_configure(path = app_path)

  # confirm the changes
  deepgs_oauth_app()
}

# restore original auth config
deepgs_auth_configure(app = original_app, api_key = original_api_key)
}
\seealso{
Other auth functions: 
\code{\link{deepgs_auth}()},
\code{\link{deepgs_deauth}()}
}
\concept{auth functions}
