% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{deepgs_deauth}
\alias{deepgs_deauth}
\title{Suspend authorization}
\usage{
deepgs_deauth()
}
\description{
Put deepgsheets4 into a de-authorized state. Instead of sending a token,
deepgsheets4 will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
deepgsheets4 in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. The user can configure their own API key
via \code{\link[=deepgs_auth_configure]{deepgs_auth_configure()}} and retrieve that key via
\code{\link[=deepgs_api_key]{deepgs_api_key()}}.
In the absence of a user-configured key, a built-in default key is used.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
deepgs_deauth()
deepgs_user()

# get metadata on the public 'deaths' spreadsheet
deepgs_example("deaths") \%>\%
  deepgs_get()
\dontshow{\}) # examplesIf}
}
\seealso{
Other auth functions: 
\code{\link{deepgs_auth_configure}()},
\code{\link{deepgs_auth}()}
}
\concept{auth functions}
