% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{deepgs_has_token}
\alias{deepgs_has_token}
\title{Is there a token on hand?}
\usage{
deepgs_has_token()
}
\value{
Logical.
}
\description{
Reports whether deepgsheets4 has stored a token, ready for use in downstream
requests.
}
\examples{
deepgs_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{deepgs_token}()}
}
\concept{low-level API functions}
