% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listinize.R
\name{deepgs_listinize}
\alias{deepgs_listinize}
\alias{deepgs_listinize.default}
\alias{deepgs_listinize.deepgsheets4Req}
\alias{deepgs_listinize.deepgsheets4Obj}
\alias{deepgs_listinize.TextFormat}
\alias{deepgs_listinize.ColorStyle}
\alias{deepgs_listinize.ChartData}
\alias{deepgs_listinize.BasicChartAxis}
\alias{deepgs_listinize.Borders}
\alias{deepgs_listinize.CellFormat}
\alias{deepgs_listinize.ExtendedValue}
\alias{deepgs_listinize.EmbeddedChart}
\alias{deepgs_listinize.DimensionProperties}
\alias{deepgs_listinize.RowData}
\alias{deepgs_listinize.ConditionValue}
\title{Transform \code{deepgsheets4} object to list}
\usage{
deepgs_listinize(x, ...)

\method{deepgs_listinize}{default}(x, ...)

\method{deepgs_listinize}{deepgsheets4Req}(x, ...)

\method{deepgs_listinize}{deepgsheets4Obj}(x, ...)

\method{deepgs_listinize}{TextFormat}(x, ...)

\method{deepgs_listinize}{ColorStyle}(x, ...)

\method{deepgs_listinize}{ChartData}(x, ...)

\method{deepgs_listinize}{BasicChartAxis}(x, ...)

\method{deepgs_listinize}{Borders}(x, ...)

\method{deepgs_listinize}{CellFormat}(x, ...)

\method{deepgs_listinize}{ExtendedValue}(x, ...)

\method{deepgs_listinize}{EmbeddedChart}(x, ...)

\method{deepgs_listinize}{DimensionProperties}(x, ...)

\method{deepgs_listinize}{RowData}(x, ...)

\method{deepgs_listinize}{ConditionValue}(x, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{list}
}
\description{
Coerce objects to lists readable by googlesheets API
}
\details{
GoogleSheets API v4 requires every object to be represented by list. These
lists are deeply nested, and as such are not very \code{R} friendly - to skim
through them, modify them etc.

Methods to \code{\link[=deepgs_listinize]{deepgs_listinize()}} provided within the package are called just before
sending request, allowing more user-friendly representation in \code{R} and
readibility on GoogleSheets end.

Listinized objects can then be parsed back into its source object using
correct \code{gen_ObjectClass()} constructor, which are used mainly for
interpreting reads from GoogleSheets API, but are also exported.
}
