% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varia.R
\name{deepgs_serial_number}
\alias{deepgs_serial_number}
\alias{as.character.deepgs_serial_number}
\alias{as.POSIXct.deepgs_serial_number}
\alias{as.POSIXlt.deepgs_serial_number}
\alias{as.Date.deepgs_serial_number}
\alias{is.deepgs_serial_number}
\title{Coerce Dates to googlesheets Serial Number}
\usage{
deepgs_serial_number(date)

\method{as.character}{deepgs_serial_number}(x, format = "", ...)

\method{as.POSIXct}{deepgs_serial_number}(x, tz = "", ...)

\method{as.POSIXlt}{deepgs_serial_number}(x, tz = "", ...)

\method{as.Date}{deepgs_serial_number}(x, ...)

is.deepgs_serial_number(x)
}
\arguments{
\item{date}{value to convert. It needs to be a value recognized by
\code{\link[lubridate:as_date]{lubridate::as_datetime()}}}

\item{x}{object of class \code{deepgs_serial_number} to coerce}

\item{format}{A character string.  The default for the \code{format}
    methods is
    \code{"\%Y-\%m-\%d \%H:\%M:\%S"} if any element has a time
    component which is not midnight, and \code{"\%Y-\%m-\%d"}
    otherwise.  If \code{\link[base]{options}("digits.secs")} is set, up to
    the specified number of digits will be printed for seconds.}

\item{...}{Further arguments to be passed from or to other methods.}

\item{tz}{A character string specifying the time zone to be used for
    the conversion.  System-specific (see \code{\link[base]{as.POSIXlt}}), but
    \code{""} is the current time zone, and \code{"GMT"} is UTC.
    Invalid values are most commonly treated as UTC, on some platforms with
    a warning.}
}
\description{
Coerce Dates to googlesheets Serial Number
}
\details{
"Serial number" format, as popularized by Lotus 1-2-3. The whole number
portion of the value (left of the decimal) counts the days since December
30th 1899. The fractional portion (right of the decimal) counts the time
as a fraction of the day. For example, January 1st 1900 at noon would be
\code{2.5}, \code{2} because it's 2 days after December 30th 1899, and \code{.5} because
noon is half a day. February 1st 1900 at 3pm would be \code{33.625}.
}
