% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{deepgs_token}
\alias{deepgs_token}
\title{Produce configured token}
\usage{
deepgs_token()
}
\value{
A \code{request} object (an S3 class provided by \link[httr:httr-package]{httr}).
}
\description{
For internal use or for those programming around the Sheets API.
Returns a token pre-processed with \code{\link[httr:config]{httr::config()}}. Most users
do not need to handle tokens "by hand" or, even if they need some
control, \code{\link[=deepgs_auth]{deepgs_auth()}} is what they need. If there is no current
token, \code{\link[=deepgs_auth]{deepgs_auth()}} is called to either load from cache or
initiate OAuth2.0 flow.
If auth has been deactivated via \code{\link[=deepgs_deauth]{deepgs_deauth()}}, \code{deepgs_token()}
returns \code{NULL}.
}
\examples{
\dontshow{if (deepgs_has_token()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
req <- request_generate(
  "sheets.spreadsheets.get",
  list(spreadsheetId = "abc"),
  token = deepgs_token()
)
req
\dontshow{\}) # examplesIf}
}
\seealso{
Other low-level API functions: 
\code{\link{deepgs_has_token}()}
}
\concept{low-level API functions}
