% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{deepgsheets4Reply}
\alias{deepgsheets4Reply}
\alias{is.deepgsheets4Reply}
\title{Reply from googlesheets API}
\usage{
deepgsheets4Reply(reply, sheetId = NULL)

is.deepgsheets4Reply(x)
}
\arguments{
\item{reply}{list containing response from googlesheets4 API}

\item{sheetId}{optional sheetId to input}

\item{x}{any R object}
}
\description{
Mostly internal post-processing function, applied after
\link{deepgsheets4 Request sending} to construct complete \link{deepgsheets4Obj} objects
}
