% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{deepgsheets4Reply_batchUpdate}
\alias{deepgsheets4Reply_batchUpdate}
\title{Reply from googlesheets API}
\usage{
deepgsheets4Reply_batchUpdate(reply, sheetId = NULL)
}
\arguments{
\item{reply}{list containing response from googlesheets4 API}

\item{sheetId}{optional sheetId to input}
}
\description{
Mostly internal post-processing function, applied after
\link{deepgsheets4 Request sending} to construct complete \link{deepgsheets4Obj} objects
}
