% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetadata.R
\name{DeveloperMetadata}
\alias{DeveloperMetadata}
\alias{is.DeveloperMetadata}
\title{Developer Metadata}
\usage{
DeveloperMetadata(
  location,
  metadataKey,
  metadataId = NULL,
  metadataValue = NULL,
  visibility = c("DOCUMENT", "PROJECT")
)

is.DeveloperMetadata(x)
}
\arguments{
\item{location}{object of class \link{DeveloperMetadataLocation}. Binds a metadata
to element in the spreadsheet}

\item{metadataKey}{string used for identification. It doesn't have to be unique}

\item{metadataId}{positive, unique integer that identifies this metadata.
If not provided during creation, random number will be created}

\item{metadataValue}{Data associated with the metadata's key.}

\item{visibility}{The metadata visibility.}

\item{x}{any R object}
}
\description{
Developer metadata associated with a location or object in a spreadsheet.
Developer metadata may be used to associate arbitrary data with various parts
of a spreadsheet and will remain associated at those locations as they move
around and the spreadsheet is edited.

If the associated object is deleted its metadata is deleted too.
}
\details{
\subsection{Metadata visibility:}{
\itemize{
\item \strong{DOCUMENT}: Document-visible metadata is accessible from any developer
project with access to the document.
\item \strong{PROJECT}: Project-visible metadata is only visible to and accessible
by the developer project that created the metadata.
}
}
}
