% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetadata.R
\name{DeveloperMetadataLookup}
\alias{DeveloperMetadataLookup}
\alias{is.DeveloperMetadataLookup}
\title{Developer Metadata Lookup}
\usage{
DeveloperMetadataLookup(
  locationType = NULL,
  metadataLocation = NULL,
  locationMatchingStrategy = NULL,
  metadataId = NULL,
  metadataKey = NULL,
  metadataValue = NULL,
  visibility = NULL
)

is.DeveloperMetadataLookup(x)
}
\arguments{
\item{locationType}{The type of location to search for.
Can be one of: \code{"ROW"}, \code{"COLUMN"}, \code{"SHEET"} or \code{"SPREADSHEET"}}

\item{metadataLocation}{Object of class \link{DeveloperMetadataLocation}. Search
for metadata on specified location.}

\item{locationMatchingStrategy}{Determines the strategy for metadata search.
Can be either \code{"EXACT_LOCATION"} or \code{"INTERSECTING_LOCATION"}}

\item{metadataId}{search for metadata of given unique metadataId}

\item{metadataKey}{search for metadata with given metadataKey}

\item{metadataValue}{search for metadata with given metadataValue}

\item{visibility}{search for metadata with given visibility. May be either
\code{"DOCUMENT"} or \code{"PROJECT"}}

\item{x}{any R object}
}
\description{
Selects DeveloperMetadata that matches all of the specified fields. Used in
\code{\link[=request_metadata_search]{request_metadata_search()}}. Any field or combination of fields may be specified.
}
\details{
If only a metadata ID is specified this considers the DeveloperMetadata with
that particular unique ID. If a metadata key is specified, this considers
all developer metadata with that key. If a key, visibility, and location
type are all specified, this considers all developer metadata with that key
and visibility that are associated with a location of that type. In general,
this selects all \link{DeveloperMetadata} that matches the intersection of all
the specified fields
}
