% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetatadataRequest.R
\name{DeveloperMetadataRequests}
\alias{DeveloperMetadataRequests}
\alias{CreateDeveloperMetadataRequest}
\alias{AppendCellsRequest}
\alias{UpdateDeveloperMetadataRequest}
\alias{DeleteDeveloperMetadataRequest}
\title{Requests to create, update or delete DeveloperMetadata}
\usage{
CreateDeveloperMetadataRequest(developerMetadata)

UpdateDeveloperMetadataRequest(dataFilters, developerMetadata, fields = NULL)

DeleteDeveloperMetadataRequest(dataFilter)
}
\arguments{
\item{developerMetadata}{object of class \link{DeveloperMetadata} containing
properties to update}

\item{dataFilters}{object of class \link{DataFilter} or list of such objects,
specifying \link{DeveloperMetadata} to update}

\item{fields}{character vector with fields to update in \emph{FieldMask} format.
Check vaild fields using \code{\link[=valid_update_fields]{valid_update_fields()}}}

\item{dataFilter}{object of class \link{DataFilter} specifying \link{DeveloperMetadata}
to remove}
}
\value{
dgs4Req object
}
\description{
Create \code{dgs4Req} objects that allow creation, update or deletion
of \link{DeveloperMetadata} in a spreadsheet. Send created requests with
\code{\link[=request_ss_batchUpdate]{request_ss_batchUpdate()}}
}
\section{CreateDeveloperMetadata}{

A request to create developer metadata.
}

\section{UpdateDeveloperMetadata}{

A request to update properties of developer metadata. Updates the properties
of the developer metadata selected by the filters to the values provided
in the \link{DeveloperMetadata} resource. Callers must specify the properties they
wish to update in the \code{fields} parameter, as well as specify at least one
\link{DataFilter} matching the metadata they wish to update.
}

\section{NA}{

A request to create developer metadata.
}

\seealso{
Other dgs4Req constructors: 
\code{\link{CellRequests}},
\code{\link{ChartRequests}},
\code{\link{DimensionRequests}},
\code{\link{PasteRequests}},
\code{\link{SheetRequests}}
}
\concept{dgs4Req constructors}
