% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dgs4_deauth}
\alias{dgs4_deauth}
\title{Suspend authorization}
\usage{
dgs4_deauth()
}
\description{
Put deepgs4 into a de-authorized state. Instead of sending a token,
deepgs4 will send an API key. This can be used to access public
resources for which no Google sign-in is required. This is handy for using
deepgs4 in a non-interactive setting to make requests that do not
require a token. It will prevent the attempt to obtain a token
interactively in the browser. The user can configure their own API key
via \code{\link[=dgs4_auth_configure]{dgs4_auth_configure()}} and retrieve that key via
\code{\link[=dgs4_api_key]{dgs4_api_key()}}.
In the absence of a user-configured key, a built-in default key is used.
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
dgs4_deauth()
dgs4_user()

# get metadata on the public 'deaths' spreadsheet
dgs4_example("deaths") \%>\%
  dgs4_get()
\dontshow{\}) # examplesIf}
}
\seealso{
Other auth functions: 
\code{\link{dgs4_auth_configure}()},
\code{\link{dgs4_auth}()}
}
\concept{auth functions}
