% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dgs4_has_token}
\alias{dgs4_has_token}
\title{Is there a token on hand?}
\usage{
dgs4_has_token()
}
\value{
Logical.
}
\description{
Reports whether deepgs4 has stored a token, ready for use in downstream
requests.
}
\examples{
dgs4_has_token()
}
\seealso{
Other low-level API functions: 
\code{\link{dgs4_token}()}
}
\concept{low-level API functions}
