% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listinize.R
\name{dgs4_listinize}
\alias{dgs4_listinize}
\alias{dgs4_listinize.default}
\alias{dgs4_listinize.deepgsheets4Req}
\alias{dgs4_listinize.deepgsheets4Obj}
\alias{dgs4_listinize.TextFormat}
\alias{dgs4_listinize.ColorStyle}
\alias{dgs4_listinize.ChartData}
\alias{dgs4_listinize.BasicChartAxis}
\alias{dgs4_listinize.ChartSpec}
\alias{dgs4_listinize.CellFormat}
\alias{dgs4_listinize.ExtendedValue}
\alias{dgs4_listinize.EmbeddedChart}
\alias{dgs4_listinize.DimensionProperties}
\alias{dgs4_listinize.RowData}
\alias{dgs4_listinize.ConditionValue}
\title{Transform \code{deepgsheets4} object to list}
\usage{
dgs4_listinize(x, ...)

\method{dgs4_listinize}{default}(x, ...)

\method{dgs4_listinize}{deepgsheets4Req}(x, ...)

\method{dgs4_listinize}{deepgsheets4Obj}(x, ...)

\method{dgs4_listinize}{TextFormat}(x, ...)

\method{dgs4_listinize}{ColorStyle}(x, ...)

\method{dgs4_listinize}{ChartData}(x, ...)

\method{dgs4_listinize}{BasicChartAxis}(x, ...)

\method{dgs4_listinize}{ChartSpec}(x, ...)

\method{dgs4_listinize}{CellFormat}(x, ...)

\method{dgs4_listinize}{ExtendedValue}(x, ...)

\method{dgs4_listinize}{EmbeddedChart}(x, ...)

\method{dgs4_listinize}{DimensionProperties}(x, ...)

\method{dgs4_listinize}{RowData}(x, ...)

\method{dgs4_listinize}{ConditionValue}(x, ...)
}
\arguments{
\item{x}{object to coerce}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{list}
}
\description{
Coerce objects to lists readable by googlesheets API
}
\details{
GoogleSheets API v4 requires every object to be represented by list. These
lists are deeply nested, and as such are not very \code{R} friendly - to skim
through them, modify them etc.

Methods to \code{dgs4_listinize} provided within the package are called just before
sending request, allowing more user-friendly representation in \code{R} and
readibility on GoogleSheets end.

Listinized objects can then be parsed back into its source object using
\code{\link[=gen_deepgsheets4Obj]{gen_deepgsheets4Obj()}} constructor, which are used mainly for
interpreting reads from GoogleSheets API, but are also exported.
}
