% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension.R
\name{DimensionGroup}
\alias{DimensionGroup}
\alias{is.DimensionGroup}
\title{Dimension Group}
\usage{
DimensionGroup(range, depth = NULL, collapsed = NULL)

is.DimensionGroup(x)
}
\arguments{
\item{range}{object of class \link{DimensionRange} that declares the range over
which this group exists.}

\item{depth}{the depth of the group - integer stating how many groups have a
range that wholly contains the range of this group.}

\item{collapsed}{This field is true if this group is collapsed. A collapsed
group remains collapsed if an overlapping group at a shallower depth is expanded.

A \code{TRUE} value does not imply that all dimensions within the group are hidden,
since a dimension's visibility can change independently from this group
property. However, when this property is updated, all dimensions within it
are set to hidden if this field is \code{TRUE}, or set to visible if this field is \code{FALSE}.}

\item{x}{any R object}
}
\description{
A group over an interval of rows or columns on a sheet, which
can contain or be contained within other groups. A group can be collapsed
or expanded as a unit on the sheet.
}
