% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension.R
\name{DimensionProperties}
\alias{DimensionProperties}
\alias{is.DimensionProperties}
\title{Properties about dimension}
\usage{
DimensionProperties(
  pixelSize = NULL,
  developerMetadata = NULL,
  hiddenByUser = NULL,
  hiddenByFilter = NULL,
  dataSourceColumnReference = NULL
)

is.DimensionProperties(x)
}
\arguments{
\item{pixelSize}{Height (if row) or width (if column) of the dimension in
pixels}

\item{developerMetadata}{object of class \link{DeveloperMetadata} or list of such
objects. Developer metadata associated with single row or column}

\item{hiddenByUser}{\code{TRUE} if this dimension is explicitly hidden}

\item{hiddenByFilter}{\strong{READ ONLY} \code{TRUE} if this dimension is being filtered}

\item{dataSourceColumnReference}{\strong{READ ONLY} if this is a column in a
data source sheet, the name of the column.}

\item{x}{any R object}
}
\description{
Describe a properties of row or column in a sheet
}
