% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtectedRange.R
\name{Editors}
\alias{Editors}
\alias{is.Editors}
\title{Editors}
\usage{
Editors(users = NULL, groups = NULL, domainUsersCanEdit = NULL)

is.Editors(x)
}
\arguments{
\item{users}{The email addresses of users with edit access to the
protected range.}

\item{groups}{The email addresses of groups with edit access to the
protected range.}

\item{domainUsersCanEdit}{\code{TRUE} if anyone in the document's domain has edit
access to the protected range. Domain protection is only supported on
documents within a domain.}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{Editors}
}
\description{
Object declaring the editors of \link{ProtectedRange}
}
