% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{EmbeddedChart}
\alias{EmbeddedChart}
\alias{is.EmbeddedChart}
\title{EmbeddedChart}
\usage{
EmbeddedChart(spec, position, borderColor = NULL, chartId = NULL)

is.EmbeddedChart(x)
}
\arguments{
\item{spec}{object of \link{ChartSpec} class, declaring chart specification}

\item{position}{object of \link{EmbeddedObjectPosition} class, declaring the
position of chart}

\item{borderColor}{object of \link{ColorStyle} class, declaring the color of
borders of the chart}

\item{chartId}{ID of the chart unique in a spreadsheet. If not set during
write, the one is chosen for you.}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{EmbeddedChart}
}
\description{
Specification of a chart embedded in a sheet
}
\seealso{
Other Chart objects constructors: 
\code{\link{ChartAxisViewWindowOptions}()},
\code{\link{ChartData}()},
\code{\link{ChartSpec}()},
\code{\link{DataLabel}()},
\code{\link{DataSourceChartProperties}()}
}
\concept{Chart objects constructors}
