% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{EmbeddedObjectPosition}
\alias{EmbeddedObjectPosition}
\title{EmbeddedObjectPosition}
\usage{
EmbeddedObjectPosition(
  type = c("anchor", "new"),
  gridCoordinate,
  offsetXPixels = 0,
  offsetYPixels = 0,
  widthPixels = 600,
  heightPixels = 371,
  sheetId = NULL
)
}
\arguments{
\item{type}{Should it be embedded in some position on existing sheet, or on
new, separate sheet}

\item{gridCoordinate}{integers declaring the anchor cell for the chart.
Indices began with 0. Valid only if \code{type = "anchor"}}

\item{offsetXPixels, ofsetYPixels}{integers declaring amount of pixels to
offset the position from the anchor cell. Valid only if \code{type = "anchor"}}

\item{widthPixels, heightPixels}{integers declaring chart size. Valid only
if \code{type = "anchor"}}

\item{sheetId}{ID of new sheet to embed the spreadsheet to. Valid only if
\code{type = "new"}. Can be unspecified - then random ID is chosen.}
}
\value{
EmbeddedObjectPosition
}
\description{
Specification of position for embedded objects (eg. charts)
}
