% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{ErrorValue}
\alias{ErrorValue}
\alias{gen_ErrorValue}
\title{Error Value}
\usage{
ErrorValue(
  type = c("ERROR", "NULL_VALUE", "DIVIDE_BY_ZERO", "VALUE", "REF", "NAME", "NUM", "N_A",
    "LOADING"),
  message
)

gen_ErrorValue(obj)
}
\arguments{
\item{type}{Error type}

\item{message}{A message with more information about the error (in the
spreadsheet's locale).}

\item{obj}{list produced by \code{\link[=deepgs_listinize]{deepgs_listinize()}}}
}
\description{
Error value of \link{ExtendedValue} - signalizing error of the
cell value.
}
\details{
Error values and their meaning:
\itemize{
\item \code{ERROR_TYPE_UNSPECIFIED}: The default error type, do not use this.
\item \code{ERROR}: Corresponds to the #ERROR! error.
\item \code{NULL_VALUE}: Corresponds to the #NULL! error.
\item \code{DIVIDE_BY_ZERO}: Corresponds to the #DIV/0 error.
\item \code{VALUE}: Corresponds to the #VALUE! error.
\item \code{REF}: Corresponds to the #REF! error.
\item \code{NAME}: Corresponds to the #NAME? error.
\item \code{NUM}: Corresponds to the #NUM! error.
\item \code{N_A}: Corresponds to the #N/A error.
\item \code{LOADING}: Corresponds to the Loading... state.
}
}
