% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{ExtendedValue}
\alias{ExtendedValue}
\alias{is.ExtendedValue}
\alias{gen_ExtendedValue}
\title{Value of the cell}
\usage{
ExtendedValue(
  numberValue = NULL,
  stringValue = NULL,
  boolValue = NULL,
  formulaValue = NULL,
  errorValue = NULL
)

is.ExtendedValue(x)

gen_ExtendedValue(obj)
}
\arguments{
\item{numberValue}{Represents a double value. Date, Times and DateTimes
need to be coerced first to}

\item{stringValue}{String value. Leading single quotes with numbers aren't
included}

\item{boolValue}{Logical value.}

\item{formulaValue}{Represents a formula to be calculated}

\item{errorValue}{object of class \link{ErrorValue} signaling error in the cell}

\item{x}{any R object}

\item{obj}{list produced by \code{\link[=deepgs_listinize]{deepgs_listinize()}}}
}
\description{
Object holding specification of cell value in \link{CellData}. Only
one of the parameters can be specified.
}
