% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{FilterView}
\alias{FilterView}
\alias{is.FilterView}
\title{FilterView}
\usage{
FilterView(
  filterViewId = NULL,
  title = NULL,
  range = NULL,
  namedRangeId = NULL,
  sortSpecs = NULL,
  filterSpecs = NULL,
  ...
)

is.FilterView(x)
}
\arguments{
\item{filterViewId}{The ID of the filter view.}

\item{title}{The name of the filter view.}

\item{range}{The range this filter view covers.}

\item{namedRangeId}{The \link{NamedRange} this filter view is backed by, if any.}

\item{sortSpecs}{The sort order per column. Later specifications are used
when values are equal in the earlier specifications.}

\item{filterSpecs}{The filter criteria for showing/hiding values per column.}

\item{...}{for deprecated fields returned from Google Sheets API}

\item{x}{any R object}
}
\description{
A filter view.
}
