% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_dgs4Obj}
\alias{gen_dgs4Obj}
\alias{@eval}
\alias{pkg_env$obj_generators}
\alias{gen_CellFormat}
\alias{gen_Border}
\alias{gen_Borders}
\alias{gen_ExtendedValue}
\alias{gen_CellData}
\alias{gen_RowData}
\alias{gen_GridData}
\alias{gen_Sheet}
\alias{gen_SheetProperties}
\alias{gen_DataSourceSheetProperties}
\alias{gen_Spreadsheet}
\alias{gen_SpreadsheetProperties}
\alias{gen_SpreadsheetTheme}
\alias{gen_OverlayPosition}
\alias{gen_EmbeddedObjectPosition}
\alias{gen_ChartAxisViewWindowOptions}
\alias{gen_ChartData}
\alias{gen_ChartSpec}
\alias{gen_DataLabel}
\alias{gen_EmbeddedChart}
\alias{gen_DataSourceChartProperties}
\alias{gen_ChartGroupRule}
\alias{gen_BasicChartAxis}
\alias{gen_BasicChartDomain}
\alias{gen_BasicSeriesDataPointStyleOverride}
\alias{gen_BasicChartSeries}
\alias{gen_BasicChartSpec}
\alias{gen_DimensionProperties}
\alias{gen_DimensionGroup}
\alias{gen_BooleanCondition}
\alias{gen_InterpolationPoint}
\alias{gen_BooleanRule}
\alias{gen_GradientRule}
\alias{gen_ConditionalFormatRule}
\alias{gen_ProtectedRange}
\alias{gen_SortSpec}
\alias{gen_FilterCriteria}
\alias{gen_FilterSpec}
\alias{gen_BasicFilter}
\alias{gen_FilterView}
\alias{gen_DeveloperMetadata}
\alias{gen_DeveloperMetadataLocation}
\alias{gen_DeveloperMetadataLookup}
\alias{gen_DataFilter}
\alias{gen_MatchedDeveloperMetadata}
\alias{gen_DataSource}
\alias{gen_DataSourceSpec}
\alias{gen_BigQueryDataSourceSpec}
\alias{gen_DataSourceParameter}
\alias{gen_DataSourceRefreshSchedule}
\alias{gen_DataSourceColumn}
\alias{gen_TextFormat}
\alias{gen_ColorStyle}
\alias{gen_TextFormatRun}
\alias{gen_UpdateValuesResponse}
\title{Generate dgs4sObj}
\usage{
gen_dgs4Obj(obj, class, sheetId = NULL)

gen_CellFormat(obj)

gen_Border(obj)

gen_Borders(obj)

gen_ExtendedValue(obj)

gen_CellData(obj)

gen_RowData(obj)

gen_GridData(obj)

gen_Sheet(obj)

gen_SheetProperties(obj)

gen_DataSourceSheetProperties(obj)

gen_Spreadsheet(obj)

gen_SpreadsheetProperties(obj)

gen_SpreadsheetTheme(obj)

gen_OverlayPosition(obj, sheetId = NULL)

gen_EmbeddedObjectPosition(obj, sheetId = NULL)

gen_ChartAxisViewWindowOptions(obj)

gen_ChartData(obj, sheetId = NULL)

gen_ChartSpec(obj, sheetId = NULL)

gen_DataLabel(obj, sheetId = NULL)

gen_EmbeddedChart(obj, sheetId = NULL)

gen_DataSourceChartProperties(obj)

gen_ChartGroupRule(obj)

gen_BasicChartAxis(obj)

gen_BasicChartDomain(obj, sheetId = NULL)

gen_BasicSeriesDataPointStyleOverride(obj)

gen_BasicChartSeries(obj, sheetId = NULL)

gen_BasicChartSpec(obj, sheetId = NULL)

gen_DimensionProperties(obj)

gen_DimensionGroup(obj, sheetId = NULL)

gen_BooleanCondition(obj)

gen_InterpolationPoint(obj)

gen_BooleanRule(obj)

gen_GradientRule(obj)

gen_ConditionalFormatRule(obj, sheetId = NULL)

gen_ProtectedRange(obj, sheetId = NULL)

gen_SortSpec(obj)

gen_FilterCriteria(obj)

gen_FilterSpec(obj)

gen_BasicFilter(obj)

gen_FilterView(obj, sheetId = NULL)

gen_DeveloperMetadata(obj, sheetId = NULL)

gen_DeveloperMetadataLocation(obj, sheetId = NULL)

gen_DeveloperMetadataLookup(obj, sheetId = NULL)

gen_DataFilter(obj, sheetId = NULL)

gen_MatchedDeveloperMetadata(obj, sheetId = NULL)

gen_DataSource(obj)

gen_DataSourceSpec(obj)

gen_BigQueryDataSourceSpec(obj)

gen_DataSourceParameter(obj)

gen_DataSourceRefreshSchedule(obj)

gen_DataSourceColumn(obj)

gen_TextFormat(obj)

gen_ColorStyle(obj)

gen_TextFormatRun(obj)

gen_UpdateValuesResponse(obj)
}
\arguments{
\item{obj}{list generated by \code{\link[=dgs4_listinize]{dgs4_listinize()}} or list received from
googlesheets API}

\item{class}{name of the class to generate}

\item{sheetId}{optional sheetId to paste into the constructor. Necessary for
some objects returned from googlesheets API}
}
\value{
deepgs4Obj of given class
}
\description{
Data flow of the \code{deepgs4} requires all of the responses from googlesheets
API to be converted back to specific class representing this object.

Some less complex object can be generated by simply passing their nested
objects as arguments to their constructor function. Others need specific
pre-process strategy to pass prepared \code{dgs4Obj} into their constructor
calls.

\code{gen_dgs4Obj} is high-level wrapper that can handle this mechanism.
Other functions listed there are more specific generators
}
