% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_dgs4Resp}
\alias{gen_dgs4Resp}
\alias{@eval}
\alias{pkg_env$resp_generators}
\title{Generate dgs4Resp}
\usage{
gen_dgs4Resp(obj, class, sheetId = NULL)
}
\arguments{
\item{obj}{list generated by \code{\link[=dgs4_listinize]{dgs4_listinize()}} or list received from
googlesheets API}

\item{class}{name of the class to generate}

\item{sheetId}{optional sheetId to paste into the constructor. Necessary for
some objects returned from googlesheets API}
}
\value{
deepgs4Obj of given class
}
\description{
Responses of \link{request_ss_batchUpdate} returned from Sheets API are coerced
into \code{dgs4Resp} objects, allowing for dropping unnecessary nesting,
transforming returned objects into \code{dgs4Obj} classes.
}
