% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{generate_request}
\alias{generate_request}
\title{Generate deepgsheets4 request}
\usage{
generate_request(
  endpoint = c("sheets.spreadsheets.get", "sheets.spreadsheets.batchUpdate",
    "sheets.spreadsheets.create", "sheets.spreadsheets.getByDataFilter"),
  params
)
}
\arguments{
\item{endpoint}{request endpoint}

\item{params}{list of params passed to the request}
}
\description{
Generate deepgsheets4 request
}
