% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1_notation.R
\name{get_A1_not}
\alias{get_A1_not}
\alias{get_A1_not.GridCoordinate}
\alias{get_A1_not.GridRange}
\title{Get A1 notation from \code{deepgsheets4} objects}
\usage{
get_A1_not(x, strict = TRUE, sheetName = NULL, sheetVec = NULL, ...)

\method{get_A1_not}{GridCoordinate}(x, strict = TRUE, sheetName = NULL, sheetVec = NULL, ...)

\method{get_A1_not}{GridRange}(
  x,
  strict = TRUE,
  sheetName = NULL,
  sheetVec = NULL,
  split_rows = FALSE,
  split_cols = FALSE,
  ...
)
}
\arguments{
\item{x}{object holding the R1C1 notation}

\item{strict}{should strict address (eg. \verb{$A$1}) be retrieved}

\item{sheetName}{character with sheetName to append}

\item{sheetVec}{object of class \code{sheetVec} describing the sheets.
Produced by \link{SpreadSheetData} object's \code{sheets} field}

\item{...}{further arguments passed to or from other methods}

\item{split_rows, split_cols}{booleans - should the rows and columns be splitted,
before returning their \emph{A1} notation?}
}
\value{
character vector or list (row-wise) containing all cell adresses in \emph{A!} notation
}
\description{
All sheets API-viable objects hold the notation for in-sheet
location by presenting rows and cols in form of 0-based indices. For function
construction the legacy spreadsheet notation \emph{A1} is required, though. This
generic allows for retrieval of such notation from objects holding this
information.
}
