% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFormat.R
\name{GradientRule}
\alias{GradientRule}
\alias{is.GradientRule}
\title{Gradient Rule}
\usage{
GradientRule(minpoint, maxpoint, midpoint = NULL)

is.GradientRule(x)
}
\arguments{
\item{minpoint, maxpoint}{objects of class \link{InterpolationPoint} specifying the
starting and final interpolation point.}

\item{midpoint}{object of class \link{InterpolationPoint}. Optional midway interpolation
point.}

\item{x}{any R object}
}
\description{
Rule that applies a gradient color scale format, based on provided
interpolation points. Used in \code{\link[=ConditionalFormatRule]{ConditionalFormatRule()}}
}
