% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridCoordinate}
\alias{GridCoordinate}
\alias{is.GridCoordinate}
\title{GridCoordinate}
\usage{
GridCoordinate(sheetId, rowIndex, columnIndex)

is.GridCoordinate(x)
}
\arguments{
\item{sheetId}{Integer - sheet ID}

\item{rowIndex, columnIndex}{position of the cell in a grid}

\item{x}{any R object}
}
\description{
Specification of cell in a grid
}
