% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridRange}
\alias{GridRange}
\alias{is.GridRange}
\alias{print.GridRange}
\title{GridRange}
\usage{
GridRange(
  sheetId,
  startRowIndex = NULL,
  endRowIndex = NULL,
  startColumnIndex = NULL,
  endColumnIndex = NULL
)

is.GridRange(x)

\method{print}{GridRange}(gr, A1 = getOption("deepgs4.A1", FALSE), ...)
}
\arguments{
\item{sheetId}{Integer - sheet ID}

\item{startRowIndex, startColumnIndex, endRowIndex, endColumnIndex}{Zero-based
indices declaring the range on the grid. \code{start} indices are inclusive, \code{end}
indices are exclusive. Missing index means that the range is unbounded on that
side.}

\item{x}{any R object}

\item{gr}{GridRange object}

\item{A1}{if \code{TRUE} then include \emph{A1} notation. Can be set globally with
\code{options("deepgs4.A1" = TRUE)}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
Specification of grid range in spreadsheet
}
\section{Warning}{

Objects representing \code{GridRange} received from GoogleSheets API v4
don't include \code{sheetId}. It is filled during response processing with \code{sheetId}
of the sheet in which the object was located in.
}

