% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFormat.R
\name{InterpolationPoint}
\alias{InterpolationPoint}
\alias{is.InterpolationPoint}
\title{Interpolation Point}
\usage{
InterpolationPoint(
  colorStyle,
  type = c("NUMBER", "PERCENT", "PERCENTILE", "MIN", "MAX"),
  value = NULL
)

is.InterpolationPoint(x)
}
\arguments{
\item{colorStyle}{object of class \link{ColorStyle} that this point will use.}

\item{type}{How the value should be interpreted}

\item{value}{A value that this interpolation point uses, which may also be
a formula. Unused for \code{"MIN"} and \code{"MAX"} \code{type}}

\item{x}{any R object}

\item{...}{provided for compatibility with deprecated Google Sheets API
objects}
}
\description{
A single interpolation point on a gradient conditional format.
Pins a gradient color scale in accordance to the color, type and value chosen.
Used in \code{\link[=GradientRule]{GradientRule()}}
}
