% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{Interval}
\alias{Interval}
\alias{is.Interval}
\title{Interval}
\usage{
Interval(startTime = NULL, endTime = NULL)

is.Interval(x)
}
\arguments{
\item{startTime}{Optional. Inclusive start of the interval. If specified, a
Timestamp matching this interval will have to be the same or after the start.}

\item{endTime}{Optional. Exclusive end of the interval.
If specified, a Timestamp matching this interval will have to be before the end.}

\item{x}{any R object}
}
\description{
Represents a time interval, encoded as a Timestamp start (inclusive)
and a Timestamp end (exclusive).

The start must be less than or equal to the end. When the start equals the
end, the interval is empty (matches no time). When both start and end are
unspecified, the interval matches any time.
}
