% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{is.deepgsheets4Req}
\alias{is.deepgsheets4Req}
\title{Requests that are to be sent to googlesheets API}
\usage{
is.deepgsheets4Req(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Requests that are to be sent to googlesheets API
}
\details{
All requests that can be sent to googlesheets API by \code{deepgsheets4}
package need to be parsed into object of \code{deepgsheets4Req}
}
\seealso{
Other deepgsheets4Req: 
\code{\link{AddChartRequest}()},
\code{\link{UpdateChartSpecRequest}()}
}
\concept{deepgsheets4Req}
