% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spreadsheet.R
\name{IterativeCalculationSettings}
\alias{IterativeCalculationSettings}
\alias{is.IterativeCalculationSettings}
\title{Settings of iterative calculations}
\usage{
IterativeCalculationSettings(maxIterations, convergenceThreshold = NULL)

is.IterativeCalculationSettings(x)
}
\arguments{
\item{maxIterations}{When iterative calculation is enabled, the maximum
number of calculation rounds to perform.}

\item{convergenceThreshold}{When iterative calculation is enabled and
successive results differ by less than this threshold value, the calculation
rounds stop.}

\item{x}{any R object}
}
\description{
Settings to control how circular dependencies are resolved with
iterative calculation.
}
