% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varia.R
\name{LineStyle}
\alias{LineStyle}
\alias{gen_LineStyle}
\title{Chart line style specification}
\usage{
LineStyle(
  width,
  type = c("SOLID", "DOTTED", "MEDIUM_DASHED", "MEDIUM_DASHED_DOTTED", "LONG_DASHED",
    "LONG_DASHED_DOTTED", "INVISIBLE", "CUSTOM")
)

gen_LineStyle(obj)
}
\arguments{
\item{width}{width of the line in pixels}

\item{type}{type of the line. For description of each type, check details section.}

\item{obj}{list produced by \code{deepgs_listinize()}}
}
\description{
Object allowing for specifying the chart line style.
}
\details{
Valid line types and their description:
\itemize{
\item \emph{SOLID}: A solid line.
\item \emph{DOTTED}: A dotted line.
\item \emph{MEDIUM_DASHED}: A dashed line where the dashes have "medium" length.
\item \emph{MEDIUM_DASHED_DOTTED}: A line that alternates between a "medium" dash and a dot.
\item \emph{LONG_DASHED}: A dashed line where the dashes have "long" length.
\item \emph{LONG_DASHED_DOTTED}: A line that alternates between a "long" dash and a dot.
\item \emph{INVISIBLE}: No dash type, which is equivalent to a non-visible line.
\item \emph{CUSTOM}: A custom dash for a line. Modifying the exact custom dash style is currently unsupported.
}
}
