% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/A1_notation.R
\name{manipulate_GridRange}
\alias{manipulate_GridRange}
\alias{split_GridRange}
\alias{extract_GridCoordinates}
\title{Manipulate GridRange object}
\usage{
split_GridRange(gr, split = c("row", "col"))

extract_GridCoordinates(gr)
}
\arguments{
\item{gr}{\link{GridRange}` object}

\item{split}{how the split be made. With \code{split = cell} it return nested list of
\link{GridCoordinate} objects: outer list containing rows, inner list: cells in a range.
With other \code{split} the function returns list of \link{GridRange} objects splitted
accordingly}
}
\description{
Functions to manipulate \link{GridRange} objects.
\itemize{
\item \code{split_GridRange} splits \code{gr} into collections of one-row or one-column
\link{GridRange} objects, depending on the \code{split} argument
\item \code{extract_GridCoordinates} extracts \link{GridCoordinate} objects specifying all
cells cantained within the provided \code{gr}
}

Split \code{GridRange} object into multiple ones: individual rows,
individual columns or individual cells
}
