% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_requests.R
\name{MergeCells}
\alias{MergeCells}
\title{MergeCells}
\usage{
MergeCells(
  gridRange,
  mergeType = c("MERGE_COLUMNS", "MERGE_ROWS", "MERGE_ALL")
)
}
\arguments{
\item{mergeType}{one of MERGE_COLUMNS, MERGE_ROWS or MERGE_ALL}

\item{GridRange}{\code{GridRange} object}
}
\value{
gsheetRequest object
}
\description{
Single MergeCells request
}
