% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CellsRequests.R
\name{MergeRequests}
\alias{MergeRequests}
\alias{MergeCellsRequest}
\alias{UnmergeCellsRequest}
\title{Requests handling cell merges}
\usage{
MergeCellsRequest(
  range,
  mergeType = c("MERGE_COLUMNS", "MERGE_ROWS", "MERGE_ALL")
)

UnmergeCellsRequest(range)
}
\arguments{
\item{range}{object of class \link{GridRange} declaring the range in which
cells need to be merged or unmerged}

\item{mergeType}{type of the merge to create}
}
\value{
deepgsheets4Req object
}
\description{
Create \code{deepgsheets4Req} objects to merge or unmerge cells in given range.
Send created requests with \code{\link[=send_batchUpdate_req]{send_batchUpdate_req()}}
}
\section{MergeCells}{

Merges all cells in the \code{range} in a way described by \code{mergeType}:
\itemize{
\item \strong{MERGE_COLUMNS} Create a merge for each column in the range
\item \strong{MERGE_ROWS} Create a merge for each row in the range
\item \strong{MERGE_ALL} Create a single merge from the range
}
}

\section{UnmergeCells}{

Unmerges cells in the given range. If the \code{range} spans multiple merges,
all will be unmerged. It must not span any merge partially
}

\seealso{
Other deepgsheets4Req constructors: 
\code{\link{DimensionGroupRequests}},
\code{\link{UpdateBordersRequest}()}
}
\concept{deepgsheets4Req constructors}
