% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{NumberFormat}
\alias{NumberFormat}
\title{Number Format}
\usage{
NumberFormat(
  type = c("TEXT", "NUMBER", "PERCENT", "CURRENCY", "DATE", "TIME", "DATE_TIME",
    "SCIENTIFIC"),
  pattern = NULL
)
}
\arguments{
\item{type}{Type of number format. See details for meaning of every field.
Required during updates.}

\item{pattern}{pattern used for formatting - if unset, default
patter of user's locale will be used if valid for given type}
}
\description{
Specification for numeric values display
}
\details{
\itemize{
\item \strong{type}:
\itemize{
\item \code{TEXT}: Text formatting, eg. \code{1000.12}
\item \code{NUMBER}: Number formatting, eg. \verb{1,000.12}
\item \code{PERCENT}: Percent formatting, eg. \verb{10.12\%}
\item \code{CURRENCY}: Currency formatting, eg. \verb{$1,000.12}
\item \code{DATE}: Date formatting, eg. \code{9/26/2008}
\item \code{TIME}: Time formatting, eg. \verb{3:59:00 PM}
\item \code{DATE_TIME}: Date+Time formatting, eg. \verb{9/26/08 15:59:00}
\item \code{SCIENTIFIC}: Scientific number formatting, eg. \code{1.01E+03}
}
\item \strong{pattern}:
Detailed information is too verbose for inclusion here. For details
refer to \href{https://developers.google.com/sheets/api/guides/formats}(GoogleSheets Date and Number Formats guide)
}
}
