% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbeddedObject.R
\name{OverlayPosition}
\alias{OverlayPosition}
\alias{is.OverlayPosition}
\title{Overlay Position}
\usage{
OverlayPosition(
  anchorCell,
  offsetXPixels = NULL,
  offsetYPixels = NULL,
  widthPixels = 600,
  heightPixels = 371
)

is.OverlayPosition(x)
}
\arguments{
\item{anchorCell}{object of class \link{GridCoordinate} - cell that the object
is anchored to}

\item{offsetXPixels, offsetYPixels}{horizontal and vertical offset in pixels
from the anchor cell}

\item{widthPixels, heightPixels}{width and height of the object in pixels}

\item{x}{any R object}
}
\description{
The location an object is overlaid on top of a grid.
}
