% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{Padding}
\alias{Padding}
\title{Cell padding}
\usage{
Padding(top, bottom, left, right)
}
\arguments{
\item{top, bottom, left, right}{Padding in pixels for each side. During update,
all fields need to be set}
}
\description{
Cell padding
}
