% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pivot.R
\name{PivotTable}
\alias{PivotTable}
\title{PivotTable}
\usage{
PivotTable(
  rows = NULL,
  columns = NULL,
  filterSpecs = NULL,
  values = NULL,
  source = NULL,
  dataSourceId = NULL,
  valueLayout = c("HORIZONTAL", "VERTICAL"),
  dataExecutionStatus = NULL
)
}
\arguments{
\item{rows, columns}{object of class \link{PivotGroup} or list of such. Each row/column
grouping in the pivot table.}

\item{filterSpecs}{object of class \link{PivotFilterSpec} or list of such. The
filters applied to the source columns before aggregating data for the pivot table.}

\item{values}{object of class \link{PivotValue} or list of such. A list of values
to include in the pivot table.}

\item{source}{Object of class \link{GridRange}. Range of values the pivot table
is reading data from, if based on \strong{GRID} sheet}

\item{dataSourceId}{The ID of the data source the pivot table is reading data
from, if based on \link{DataSource}.}

\item{valueLayout}{Whether values should be listed horizontally (as columns)
or vertically (as rows).}

\item{dataExecutionStatus}{\strong{READ ONLY} object of class \link{DataExecutionStatus}
if pivot table is based on \link{DataSource}}

\item{...}{additional args received from Googlesheets API}
}
\description{
PivotTable
}
