% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProtectedRange.R
\name{ProtectedRange}
\alias{ProtectedRange}
\alias{is.ProtectedRange}
\title{ProtectedRange}
\usage{
ProtectedRange(
  range = NULL,
  namedRangeId = NULL,
  description = NULL,
  warningOnly = NULL,
  unprotectedRanges = NULL,
  editors = NULL,
  protectedRangeId = NULL,
  requestingUserCanEdit = NULL
)

is.ProtectedRange(x)
}
\arguments{
\item{range}{object of class \link{GridRange}. The range that is being protected.
The range may be fully unbounded, in which case this is considered a protected sheet.}

\item{namedRangeId}{ID of the \link{NamedRange} this protected range is backed by.}

\item{description}{The description of this protected range.}

\item{warningOnly}{\code{TRUE} if this protected range will show a warning when editing.}

\item{unprotectedRanges}{object of class \link{GridRange} or list of such objects.
Unprotected ranges within a protected sheet (fully unbounded \code{range})}

\item{editors}{object of class \link{Editors}. The users and groups with edit
access to the protected range. This field is only visible to users with
edit access to the protected range and the document. Editors are not
supported with \code{warningOnly} protection.}

\item{protectedRangeId}{\strong{READ ONLY} The ID of the protected range.}

\item{requestingUserCanEdit}{\strong{READ ONLY} \code{TRUE} if the user who requested this protected
range can edit the protected area.}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{ProtectedRange}
}
\description{
Object declaring a protected range on a sheet, or a protected sheet (with unbounded \code{range}).
ProtectedRange can be declared either on the basis of \code{range} or \code{namedRangeId}.
If describing a protected sheet, \code{uprotectedRanges} can also be specified.
}
\section{warningOnly}{

Warning-based protection means that every user can edit data in the protected range,
except editing will prompt a warning asking the user to confirm the edit.

When writing: if \code{warningOnly} is true, then \code{editors} is ignored.
Additionally, if this field is changed from \code{TRUE} to \code{FALSE} and the
\code{editors} field is not set (nor included in the field mask), then the
\code{editors} will be set to all the editors in the document.
}

