% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{request_generate}
\alias{request_generate}
\title{Generate deepgsheets4 request}
\usage{
request_generate(endpoint, params, token = deepgs_token())
}
\arguments{
\item{endpoint}{request endpoint}

\item{params}{list of params passed to the request}

\item{token}{token to be included}
}
\description{
Generate deepgsheets4 request
}
