% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{request_make}
\alias{request_make}
\title{Make request to googlesheets API}
\usage{
request_make(x, ..., encode = "json")
}
\arguments{
\item{x}{List holding components for HTTP request}

\item{...}{Optional arguments passed to the HTTP method}

\item{encode}{how the body should be encoded}
}
\description{
Make request to googlesheets API
}
