% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetadata.R
\name{request_metadata_get}
\alias{request_metadata_get}
\title{Get specific metadata from spreadsheet}
\usage{
request_metadata_get(spreadsheetId, metadataId)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{metadataId}{unique developer metadata ID}
}
\value{
\code{DeveloperMetadata} object
}
\description{
Sends a request to receive \link{DeveloperMetadata} object
of specific \code{metadataId}
}
