% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetadata.R
\name{request_metadata_search}
\alias{request_metadata_search}
\title{Search for metadata in spreadsheet}
\usage{
request_metadata_search(spreadsheetId, dataFilters)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{dataFilters}{object of class \link{DataFilter} or list of such objects}
}
\value{
list of \code{dgs4Response} objects of class \code{MatchedDeveloperMetadata}
or \code{NULL} if nothing is found
}
\description{
Sends a request to search for \link{DeveloperMetadata} objects
matching the specified \code{dataFilters} in given spreadsheet
}
