% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_spreadsheets.R
\name{request_ss_batchUpdate}
\alias{request_ss_batchUpdate}
\title{Update spreadsheet through Sheets API}
\usage{
request_ss_batchUpdate(
  spreadsheetId,
  ...,
  includeSpreadsheetInResponse = NULL,
  fields = NULL,
  responseRanges = NULL,
  responseIncludeGridData = NULL,
  add_params = list(),
  .dots = list()
)
}
\arguments{
\item{...}{objects of class \code{dgs4Req}}

\item{includeSpreadsheetInResponse}{Determines if the update response should
include the \code{updatedSpreadsheet}}

\item{responseRanges}{One or multiple character string in \emph{A1} notation. For
more information see \code{\link[=get_A1]{get_A1()}}. Limits the ranges included in the response spreadsheet.}

\item{responseIncludeGridData}{True if grid data should be returned.
Meaningful only if \code{includeSpreadsheetInResponse} is \code{TRUE}. This parameter
is ignored if a field mask was set in the request.}

\item{add_params}{Named list of additional parameters to include in the request}

\item{.dots}{the same as \code{...}, but applied as a list}
}
\value{
Named list containing:
\itemize{
\item \code{spreadsheetId}: ID of the spreadsheet that the updates were applied to
\item \code{replies}: List of \code{dgs4Resp} objects or empty lists
\item \code{updatedSpreadsheet}: Object of class \link{Spreadsheet} showing the state
of spreadsheet after the updates. Only if \code{includeSpreadsheetInResponse}
was set to \code{TRUE}.
}
}
\description{
Sends request to the \href{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/batchUpdate}{spreadsheets.batchUpdate}
method. Allows modifying existing spreadsheet using one or more \code{dgs4Req}
objects.
}
\details{
Each request is validated before being applied. If \strong{any} request is not
valid then the entire \emph{batchUpdate} request will fail and nothing will be applied.
Some requests have replies to give you some information about how they are
applied. The replies will mirror the requests. For example, if you applied
4 updates and the 3rd one had a reply, then the returned \code{replies} will be a
list containing:
\itemize{
\item 2 empty \code{list()} objects
\item one \code{dgs4Resp} object
\item 1 more empty \code{list()} object
replies, the actual reply, and another empty reply, in that order.
}
\subsection{Warning}{

Due to the collaborative nature of spreadsheets, it is not
guaranteed that the spreadsheet will reflect exactly your changes after
this completes, however it is guaranteed that the updates in the request
will be applied together atomically. Your changes may be altered with
respect to collaborator changes. If there are no collaborators, the
spreadsheet should reflect your changes.
}
}
\seealso{
Other Sheets Spreadsheets requests: 
\code{\link{request_ss_create}()},
\code{\link{request_ss_get}()}
}
\concept{Sheets Spreadsheets requests}
