% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_values.R
\name{request_ss_clear_values}
\alias{request_ss_clear_values}
\title{Clear values in a given range from sheet}
\usage{
request_ss_clear_values(spreadsheetId, range)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{range}{specified range in \emph{A1} or \emph{R1C1} notation.
See \code{\link[=get_A1_not]{get_A1_not()}} for more info.}
}
\value{
a list containing \code{spreadsheetId} and \code{clearedRange}
}
\description{
Send a request to clear values from cells at given range.
Sends request to \href{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/clear}{spreadsheets.values.clear}
method.
}
\seealso{
Other Sheets Values requests: 
\code{\link{request_ss_append_values}()},
\code{\link{request_ss_get_values}()},
\code{\link{request_ss_update_values}()}
}
\concept{Sheets Values requests}
