% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_values.R
\name{request_ss_get_values}
\alias{request_ss_get_values}
\title{Get values in a given range from spreadsheet}
\usage{
request_ss_get_values(
  spreadsheetId,
  range,
  majorDimension = c("ROWS", "COLUMNS"),
  valueRenderOption = c("FORMATTED_VALUE", "UNFORMATTED_VALUE", "FORMULA"),
  dateTimeRenderOption = c("SERIAL_NUMBER", "FORMATTED_STRING"),
  add_params = list()
)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{range}{specified range in \emph{A1} or \emph{R1C1} notation.
See \link{get_A1_not} for more info.}

\item{majorDimension}{The major dimension that results should use.}

\item{dateTimeRenderOption}{How values should be represented in the output.}

\item{add_params}{named list of additional params to include in HTML request}

\item{valueRendeOption}{How values should be represented in the output.}
}
\value{
\link{ValueRange} object
}
\description{
Sends request to the \href{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets.values/get}{spreadsheets.values.get}
method.
}
\section{majorDimension}{

If the spreadsheet data on Sheet1 are:\tabular{lll}{
    \tab A \tab B \cr
   1 \tab 1 \tab 2 \cr
   2 \tab 3 \tab 4 \cr
}


Then \code{majorDimension = "ROWS"} returns values in form of:
\code{list(list(1, 2), list(3, 4))} and \code{majorDimension = "COLUMNS"}:
\code{list(list(1, 3), list(2, 4))}.
}

\section{valueRenderOption}{

\itemize{
\item \strong{FORMATTED_VALUE}: Values will be calculated & formatted in the reply
according to the cell's formatting. Formatting is based on the
spreadsheet's locale, not the requesting user's locale. For example,
if A1 is 1.23 and A2 is =A1 and formatted as currency,
then A2 would return "$1.23".
\item \strong{UNFORMATTED_VALUE}: Values will be calculated, but not formatted in the
reply. For example, if A1 is 1.23 and A2 is =A1 and formatted as currency,
then A2 would return the number 1.23.
\item \strong{FORMULA}: Values will not be calculated. The reply will include the
formulas. For example, if A1 is 1.23 and A2 is =A1 and formatted as
currency, then A2 would return "=A1".
}
}

\section{dateTimeRenderOption}{

\itemize{
\item \strong{SERIAL_NUMBER}: Instructs date, time, datetime, and duration fields to
be output as doubles in "serial number" format. For more info, see \code{\link[=dgs4_serial_number]{dgs4_serial_number()}}
\item \strong{FORMATTED_STRING}: Instructs date, time, datetime, and duration fields
to be output as strings in their given number format (which depends
on the spreadsheet locale).
}
}

\seealso{
Other Sheets Values requests: 
\code{\link{request_ss_append_values}()},
\code{\link{request_ss_clear_values}()},
\code{\link{request_ss_update_values}()}
}
\concept{Sheets Values requests}
