% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DeveloperMetadata.R
\name{request_ss_metadata}
\alias{request_ss_metadata}
\alias{request_ss_metadata_get}
\alias{request_ss_metadata_search}
\title{Send Google Sheets \emph{spreadsheets.metadata} requests}
\usage{
request_ss_metadata_get(spreadsheetId, metadataId)

request_ss_metadata_search(spreadsheetId, dataFilters)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{metadataId}{unique developer metadata ID}

\item{dataFilters}{object of class \link{DataFilter} or list of such objects}
}
\description{
These functions access the \emph{spreadsheets.metadata} Sheets API
endpoint. They allow getting \link{DeveloperMetadata} objects embedded into
GoogleSheet spreadsheet.
}
\section{Get}{

Sends a request to receive \link{DeveloperMetadata} object of specific \code{metadataId}
from spreadsheet of specific \code{spreadsheetId}.
\subsection{Returns:}{

\link{DeveloperMetadata} object
}
}

\section{Search}{

Sends a request to search for \link{DeveloperMetadata} objects
matching the specified \code{dataFilters} in given spreadsheet
\subsection{Returns:}{

list of \code{dgs4Resp} objects of class \code{MatchedDeveloperMetadata}
or \code{NULL} if nothing is found
}
}

\concept{Sheet API requests}
