% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_sheets.R
\name{request_ss_sheet_copy}
\alias{request_ss_sheet_copy}
\title{Copy spreadsheet sheet to another spreadsheet}
\usage{
request_ss_sheet_copy(
  spreadsheetId,
  sheetId,
  destinationSpreadsheetId,
  add_params = list()
)
}
\arguments{
\item{spreadsheetId}{ID of the source spreadsheet}

\item{sheetId}{ID of the source sheet}

\item{destinationSpreadsheetId}{ID of the destination spreadsheet}

\item{add_params}{Named list of additional parameters for http method}
}
\value{
\link{SheetProperties} of copied sheet
}
\description{
Sends request to the \href{https://developers.google.com/sheets/api/reference/rest/v4/spreadsheets/get}{spreadsheets.sheets.copyTo}
method, allowing copying specified sheet from one spreadsheet to another.
}
