% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/req_batchUpdate.R
\name{send_batchUpdate_req}
\alias{send_batchUpdate_req}
\alias{is.deepgsheets4Req}
\alias{is.deepgsheets4Response}
\title{Send \emph{batchUpdate} request to googlesheets API}
\usage{
send_batchUpdate_req(spreadsheetId, ..., .dots = list())

is.deepgsheets4Req(x)

is.deepgsheets4Response(x)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet to apply the changes}

\item{...}{objects of class \code{deepgsheets4Req}}

\item{.dots}{the same as \code{...}, but applied as a list}

\item{x}{any R object}
}
\value{
list containing \code{spreadsheetId} and \code{replies}: list of \code{deepgsheets4Response}
objects
}
\description{
This function sends \emph{batchUpdate} request to googlesheets API
based on multiple \code{deepsheets4Req} objects
}
