% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_values_request.R
\name{send_clear_values_req}
\alias{send_clear_values_req}
\title{Clear values in a given range from spreadsheet}
\usage{
send_clear_values_req(spreadsheetId, range)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{range}{specified range in \emph{A1} or \emph{R1C1} notation.
See \link{get_A1_not} for more info.}
}
\value{
a list containing \code{spreadsheetId} and \code{clearedRange}
}
\description{
Send a request to clear values from cells at given range
}
\seealso{
Other Sheets Values requests: 
\code{\link{send_get_values_req}()}
}
\concept{Sheets Values requests}
