% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{send_create_req}
\alias{send_create_req}
\title{Send \emph{create} request to googlesheets API}
\usage{
send_create_req(spreadsheet)
}
\arguments{
\item{spreadsheet}{object of class \link{Spreadsheet}}
}
\value{
response containing created object of class \link{Spreadsheet}
}
\description{
This function sends \emph{create} request to googlesheets API
to create specified spreadsheet
}
