% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sheet.R
\name{Sheet}
\alias{Sheet}
\alias{is.Sheet}
\title{Googlesheets Sheet specification}
\usage{
Sheet(
  properties = NULL,
  data = NULL,
  merges = NULL,
  conditionalFormats = NULL,
  charts = NULL,
  developerMetadata = NULL
)

is.Sheet(x)
}
\arguments{
\item{properties}{object of class \link{SheetProperties}, describing the
properties of the sheet}

\item{data}{object of class \link{GridData} or list of such. Data present in the sheet.}

\item{merges}{object of class \link{GridRange} or list of such. Describe merged
cells.}

\item{charts}{object of class \link{EmbeddedChart} or list of such. Charts embedded
on a sheet}

\item{x}{any R object}

\item{conditionalFormat}{object of class \link{ConditionalFormatRule} or list of such.
Describe conditional format rules applied to the cells in the sheet}

\item{rowGroups, columnGroups}{objects of class \link{DimensionGroup} or list
of such, describing grouped rows and columns}
}
\description{
Specification of singular Sheet of the google spreadsheet
}
