% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetsProperties.R
\name{SheetProperties}
\alias{SheetProperties}
\alias{is.SheetProperties}
\title{SheetProperties}
\usage{
SheetProperties(
  sheetId,
  title,
  index = NULL,
  sheetType = c("GRID", "OBJECT", "DATA_SOURCE"),
  gridProperties = NULL,
  hidden = NULL,
  tabColorStyle = NULL,
  rightToLeft = NULL,
  dataSourceSheetProperties = NULL
)

is.SheetProperties(x)
}
\arguments{
\item{sheetId}{ID of the sheet}

\item{title}{Name of the sheet}

\item{index}{Index of the sheet. If not set during addition, new sheet will
be inserted at the end of the sheets}

\item{sheetType}{Type of the spreadsheet}

\item{gridProperties}{object of class \link{GridProperties}. If not set, then I don't know}

\item{hidden}{boolean indicating if sheet should be hidden in the UI}

\item{tabColorStyle}{object of class \link{ColorStyle}, describing color of the tab}

\item{rightToLeft}{boolean indicating if sheet should be RTL}

\item{dataSourceSheetProperties}{object of \link{DataSourceSheetProperties}. Field
is only accessible during reads from Sheets API.}

\item{x}{any R object}

\item{...}{additional deprecated fields returned from GoogleSheet API}
}
\value{
SheetProperties object
}
\description{
Properties of a singular sheet in SpreadSheet
}
