% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SheetRequests.R
\name{SheetRequests}
\alias{SheetRequests}
\alias{AddSheetRequest}
\alias{UpdateSheetPropertiesRequest}
\alias{DeleteSheet}
\alias{DeleteSheetRequest}
\alias{DuplicateSheetRequest}
\title{Requests to add, modify and delete Sheet}
\usage{
AddSheetRequest(properties)

UpdateSheetPropertiesRequest(properties, fields = NULL)

DeleteSheetRequest(sheetId)

DuplicateSheetRequest(
  sourceSheetId,
  insertSheetIndex,
  newSheetId = NULL,
  newSheetName = NULL
)
}
\arguments{
\item{properties}{object of class \link{SheetProperties}}

\item{fields}{which of the properties specified within \code{properties} should
be updated}

\item{sheetId}{integer representing the sheet to delete}

\item{sourceSheetId}{ID of the sheet to duplicate}

\item{insertSheetIndex}{The zero-based index where the new sheet should be
inserted. The index of all sheets after this are incremented.}

\item{newSheetId}{If set, the ID of the new sheet. If not set, an ID is chosen.
ID must not conflict with any existing sheet ID and it must be non-negative.}

\item{newSheetName}{The name of the new sheet. It should be unique amongst
spreadsheet. If empty, a new name is chosen automatically.}
}
\value{
dgs4Req object
}
\description{
Create \code{dgs4Req} objects that allow for addition, modification
and deletion of singular sheet. Send created requests with \code{\link[=request_ss_batchUpdate]{request_ss_batchUpdate()}}
}
\section{Add}{

Adds a new sheet. When a sheet is added at a given index,
all subsequent sheets' indexes are incremented. All properties are optional.
The sheetId field is optional; if one is not set, an ID will be randomly generated.
}

\section{Update}{

\code{sheetId} field of the \code{properties} will be used to identify which sheet
properties need to be updated. Only properties specified in \code{fields} will
be updated.
}

\section{Delete}{

Sheet of the given \code{sheetId} will be deleted. If the sheet is of \code{DATA_SOURCE}
type, the associated DataSource is also deleted.
}

\section{Duplicate}{

Duplicates the contents of a sheet.
If the source sheet is of DATA_SOURCE type, its backing \emph{DataSource} is also
duplicated and associated with the new copy of the sheet. No data execution
is triggered, the grid data of this sheet is also copied over but only
available after the batch request completes.
}

\seealso{
Other dgs4Req constructors: 
\code{\link{CellRequests}},
\code{\link{ChartRequests}},
\code{\link{DeveloperMetadataRequests}},
\code{\link{DimensionRequests}},
\code{\link{PasteRequests}}
}
\concept{dgs4Req constructors}
