% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Filters.R
\name{SortSpec}
\alias{SortSpec}
\alias{is.SortSpec}
\title{SortSpec}
\usage{
SortSpec(
  sortOrder = c("ASCENDING", "DESCENDING"),
  foregroundColorStyle = NULL,
  backgroundColorStyle = NULL,
  dimensionIndex = NULL,
  dataSourceColumnReference = NULL,
  ...
)

is.SortSpec(x)
}
\arguments{
\item{sortOrder}{the order in which data should be sorted.}

\item{foregroundColorStyle}{object of class \link{ColorStyle}. The foreground color
to sort by; cells with this foreground color are sorted to the top. Must be
an RGB-type color.}

\item{backgroundColorStyle}{object of class \link{ColorStyle}. The background color
to sort by; cells with this background color are sorted to the top. Must be
an RGB-type color.}

\item{dimensionIndex}{Index of dimension this sort should be applied to, if
sort is related to dimension on \strong{GRID} sheet}

\item{dataSourceColumnReference}{Unique name of the column, if sort is related
to column on \strong{DATA_SOURCE} sheet}

\item{...}{for deprecated fields returned from Sheets API}

\item{x}{any R object}
}
\description{
A sort order associated with a specific column or row. Exactly
one of \code{dimensionIndex} or \code{dataSourceColumnReference} needs to be specified.
}
