% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpreadSheetsData.R
\docType{class}
\name{SpreadSheetData}
\alias{SpreadSheetData}
\title{SpreadSheetData}
\description{
Object holding and querying for data regarding specific
spreadsheet
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{spreadsheetProperties}}{object of class spreadSheetProperties.
Acquired during object initialization}

\item{\code{sheetProperties}}{objects of class \code{sheetProperties}. Acquired during
object initialization}

\item{\code{namedRanges}}{objects of class \code{namedRange}.}

\item{\code{merges}}{objects of class \code{GridRange} describing the merged cells}

\item{\code{filterViews}}{objects filtered views}

\item{\code{protectedRanges}}{objects protected ranges}

\item{\code{basicFilter}}{objects basic filters}

\item{\code{charts}}{objects charts}

\item{\code{bandedRanges}}{banded Ranges objects}

\item{\code{developerMetadata}}{objects}

\item{\code{rowGroups}}{rowGroups objects}

\item{\code{columnGroups}}{columnGroups objects}

\item{\code{slicers}}{slicers objects}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{spreadSheetId}}{ID of the spreadsheet}

\item{\code{valid_fields}}{Fields valid to get with \code{get_data()} method}

\item{\code{sheets}}{Named vector of sheets inside this spreadsheet. Names of the
vector are the titles of sheets, and values are the sheets IDs}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SpreadSheetData-new}{\code{SpreadSheetData$new()}}
\item \href{#method-SpreadSheetData-get_data}{\code{SpreadSheetData$get_data()}}
\item \href{#method-SpreadSheetData-clone}{\code{SpreadSheetData$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpreadSheetData-new"></a>}}
\if{latex}{\out{\hypertarget{method-SpreadSheetData-new}{}}}
\subsection{Method \code{new()}}{
Initialize object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpreadSheetData$new(spreadSheetId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{spreadSheetId}}{ID ofthe spreadsheet}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpreadSheetData-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-SpreadSheetData-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get data about specific spreadsheet or sheet properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpreadSheetData$get_data(fields, refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fields}}{character vector of fields to get. Valid values can be
retrieved with \code{SpreadSheetData$valid_fields}}

\item{\code{refresh}}{boolean indicating if data should be refreshed}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SpreadSheetData-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SpreadSheetData-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SpreadSheetData$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
