% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spreadsheet.R
\name{SpreadsheetProperties}
\alias{SpreadsheetProperties}
\alias{gen_SpreadsheetProperties}
\title{SpreadsheetProperties}
\usage{
SpreadsheetProperties(
  title,
  locale,
  autoRecalc = c("ON_CHANGE", "MINUTE", "HOUR"),
  timeZone = NULL,
  defaultFormat = NULL,
  iterativeCalculationSettings = NULL,
  spreadsheetTheme = NULL,
  ...
)

gen_SpreadsheetProperties(obj)
}
\arguments{
\item{title}{Title of the spreadsheet}

\item{locale}{Locale of the spreadsheet}

\item{autoRecalc}{Amount of time to wait before recalculation of
volatile functions}

\item{timeZone}{Time zone of the spreadsheet, in \code{CLDR} format}

\item{defaultFormat}{object of \link{CellFormat} class: default format of all cells.}

\item{iterativeCalculationSettings}{object of \link{IterativeCalculationSettings}
declaring how circular references are resolved with iterative calculation}

\item{spreadsheetTheme}{object of \link{SpreadsheetTheme} class: theme applied to
spreadsheet}

\item{...}{additional deprecated fields returned from GoogleSheet API}

\item{obj}{list generated by \code{\link[=deepgs_listinize]{deepgs_listinize()}}}
}
\description{
Properties of whole spreadsheet
}
