% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varia.R
\name{TextFormat}
\alias{TextFormat}
\alias{is.TextFormat}
\title{TextFormat}
\usage{
TextFormat(
  foregroundColorStyle = NULL,
  fontFamily = NULL,
  fontSize = NULL,
  bold = NULL,
  italic = NULL,
  strikethrough = NULL,
  underline = NULL,
  link = NULL,
  ...
)

is.TextFormat(x)
}
\arguments{
\item{foregroundColorStyle}{object of class \code{ColorStyle}}

\item{fontFamily}{name of the font family}

\item{fontSize}{size of the font}

\item{bold}{boolean: should the text be bolded?}

\item{italic}{boolean: should the text be italic?}

\item{strikethrough}{boolean: should the text be strikethrough?}

\item{underline}{boolean: should the text be underlined?}

\item{link}{URL if text should link somewhere}

\item{...}{additional deprecated fields returned from GoogleSheet API}

\item{x}{any R file}
}
\description{
Class holding specification for text format. All arguments
are optional, but at least one must be present for object creation.
}
