% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataSource.R
\name{TimeOfDay}
\alias{TimeOfDay}
\alias{is.TimeOfDay}
\title{TimeOfDay}
\usage{
TimeOfDay(hours = NULL, minutes = NULL, seconds = NULL, nanos = NULL)

is.TimeOfDay(x)
}
\arguments{
\item{hours}{Hours of day in 24 hour format. Should be from 0 to 23. An API
may choose to allow the value \code{24} for scenarios like business closing time.}

\item{minutes}{Minutes of hour of day. Must be from 0 to 59.}

\item{seconds}{Seconds of minutes of the time. Must normally be from 0 to 59.
An API may allow the value 60 if it allows leap-seconds.}

\item{nanos}{Fractions of seconds in nanoseconds. Must be from 0 to 999,999,999.}

\item{x}{any R object}
}
\description{
Represents a time of day. The date and time zone are either not significant
or are specified elsewhere. An API may choose to allow leap seconds
}
