% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{to_GridData_from_df}
\alias{to_GridData_from_df}
\title{Create GridData out of \code{data.frame}}
\usage{
to_GridData_from_df(
  df,
  startRow,
  startColumn,
  names_format = NULL,
  values_format = NULL,
  names_types = NULL,
  values_types = NULL
)
}
\arguments{
\item{df}{\code{data.frame} object to use the data from}

\item{startRow, startColumn}{zero-based indices of first row and column
this \code{GridData} refers to}

\item{names_format, values_format}{object of class \link{CellFormat} that provides
format for column names and values}

\item{names_types, values_types}{vector of types to use for column names and
values. If nothing is provided, types are guessed for \code{value_types} (\code{names_types})
are always presumed to be of \code{c} type. For more info read \emph{details} section}
}
\description{
Higher-level wrapper for \link{GridData} creation, taking as input
data in form of \code{data.frame}.
}
\details{
\code{names_types} and \code{values_types} could be provided as singular value, in which
case ALL names and ALL columns will be saved as the same class, or as a vector
of the same length as number of columns, in which case the every column will
have its own class provided.

Possible values are:
\itemize{
\item \code{n} for values formatted as numeric
\item \code{c} for values formatted as character
\item \code{b} for values formatted as boolean (logical)
\item \code{dt} for values to be coerced by \code{\link[=deepgs_serial_number]{deepgs_serial_number()}}
\item \code{f} if values are characters that need to be interpreted as googlesheets
formulas. This type is never guessed.
}

If \code{values_types} argument is kept as \code{NULL}, the types will be interpreted
from data.frame column type.
}
