% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_merge.R
\name{UnmergeCellsRequest}
\alias{UnmergeCellsRequest}
\title{Unmerge cells request}
\usage{
UnmergeCellsRequest(gridRange)
}
\arguments{
\item{gridRange}{object of class \link{GridRange} declaring the cells to be unmerged}
}
\value{
deepgsheets4Req
}
\description{
Request unmerging all merged cells in given range
}
\seealso{
Other deepgsheets4Req: 
\code{\link{AddChartRequest}()},
\code{\link{MergeCellsRequest}()},
\code{\link{UpdateChartSpecRequest}()},
\code{\link{is.deepgsheets4Req}()}
}
\concept{deepgsheets4Req}
