% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_cells.R
\name{UpdateCellsRequest}
\alias{UpdateCellsRequest}
\title{Update data in cells}
\usage{
UpdateCellsRequest(
  rows,
  fields = c("userEnteredValue", "userEnteredFormat", "note", "textFormatRuns",
    "dataValidation", "pivotTable", "dataSourceTable", "dataSourceFormula"),
  start = NULL,
  range = NULL
)
}
\arguments{
\item{rows}{object of class \link{RowData} or list of multiple such objects}

\item{fields}{name of \link{CellData} fields to update}

\item{start}{object of class \link{GridCoordinate}. The coordinate to start writing
data at. Any number of rows and columns (including a different number of
columns per row) may be written. Can't be specified alongside \code{range.}}

\item{range}{object of class \link{GridRange} or list of such objects.
The range to write data to. If the data in rows does not cover the entire
requested range, the fields matching those set in fields will be cleared.
Can't be specified alongside \code{start}.}
}
\description{
Construct request to update data in specified range
}
