% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_dimensions.R
\name{UpdateDimensionPropertiesRequest}
\alias{UpdateDimensionPropertiesRequest}
\title{Request update of dimension properties}
\usage{
UpdateDimensionPropertiesRequest(
  properties,
  fields = c("pixelSize", "developerMetadata", "hiddenByUser"),
  range = NULL,
  dataSourceSheetRange = NULL
)
}
\arguments{
\item{properties}{object of class \link{DimensionProperties} providing data to
update}

\item{fields}{which fields of current properties need to be updated}

\item{range}{object of class \link{DimensionRange} specifying the range of columns
or rows to update on \code{GRID} sheets}

\item{dataSourceSheetRange}{object of class \code{DataSourceDimensionRange} specifying
the range of columns in \code{DATA_SOURCE} sheet}
}
\description{
Send a request to update properties of dimension (column/row) in
a given range. Range can be specified with either \code{range} (for \code{GRID} sheets)
or \code{dataSourceSheetRange} (for \code{DATA_SOURCE} sheets) arguments
}
