% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Values.R
\name{UpdateValuesResponse}
\alias{UpdateValuesResponse}
\title{Update Values Response}
\usage{
UpdateValuesResponse(
  spreadsheetId,
  updatedRange,
  updatedRows,
  updatedColumns,
  updatedCells,
  updatedData = NULL
)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet the updates were applied to}

\item{updatedRange}{range in \emph{A1 notation} where the updates were applied to}

\item{updatedRows}{number of rows where at least one cell in the row was updated}

\item{updatedColumns}{number of columns where at least one cell in the column
was updated}

\item{updatedCells}{number of cells updated}

\item{updatedData}{object of class \link{ValueRange} containing the values of the
cells after updates were applied. Only if the \code{includeValuesInResponse} of
\code{\link[=send_update_values_req]{send_update_values_req()}} were \code{TRUE}}
}
\value{
UpdateValuesResponse object
}
\description{
Response when updating a range of values in a spreadsheet
}
