% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_fields_validity.R
\name{valid_update_fields}
\alias{valid_update_fields}
\title{Check validity of specified fields}
\usage{
valid_update_fields(updateRequest = c("Cells"))
}
\arguments{
\item{updateRequest}{which update request to check}

\item{fields}{character vector in}
}
\value{
character vector
}
\description{
All update requests to be sent by \code{\link[=send_batchUpdate_req]{send_batchUpdate_req()}} take
a \code{fields} argument in form of \href{https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.FieldMask}{FieldMask}{target="_blank"}

\emph{FieldMask} format is helpful for specifying accurately which fields should
be updated by BatchUpdate request.

This function works as a helper for retrieval of valid update fields per
request.
}
