% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Values.R
\name{ValueRange}
\alias{ValueRange}
\title{Values within range of a spreadsheet}
\usage{
ValueRange(range, values, majorDimension = c("ROWS", "COLUMNS"))
}
\arguments{
\item{range}{The range the values cover, in \emph{A1 notation}. For output, this
range indicates the entire requested range, even though the values will
exclude trailing rows and columns. When appending values, this field
represents the range to search for a table, after which values will be appended.}

\item{values}{nested list containing values. Each nested list is one
row or column of the data (depending on \code{majorDimension} specification).
Values can be of different types, but only atomic types}

\item{majorDimension}{The major dimension of the values.}
}
\description{
Object containing values within range of a spreadsheet. Mostly
received and send through \code{spreadsheets.values} methods.
}
