% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_chart.R
\name{AddChartRequest}
\alias{AddChartRequest}
\title{AddChartRequest}
\usage{
AddChartRequest(chart)
}
\arguments{
\item{chart}{an \link{EmbeddedChart} object containing specification for a chart}
}
\value{
deepgsheets4Req
}
\description{
Constructs request for creation of googlesheets chart
}
\seealso{
Other deepgsheets4Req: 
\code{\link{UpdateChartSpecRequest}()},
\code{\link{is.deepgsheets4Req}()}
}
\concept{deepgsheets4Req}
