% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{BasicChartAxis}
\alias{BasicChartAxis}
\alias{is.BasicChartAxis}
\title{BasicChartAxis}
\usage{
BasicChartAxis(
  position = c("BOTTOM_AXIS", "LEFT_AXIS", "RIGHT_AXIS"),
  title = NULL,
  format = NULL,
  viewWindowOptions = NULL,
  horizontalAlignment = NULL
)

is.BasicChartAxis(x)
}
\arguments{
\item{position}{Position of the axis}

\item{title}{Title for the axis. If set, overrides title infered from data.}

\item{format}{object of class \link{TextFormat}. Specifies format of text.}

\item{viewWindowOptions}{object of class \link{ChartAxisViewWindowOptions}.
Specifies view on this axis}

\item{horizontalAlignment}{Alignment of the title text on the axis. Valid positions
are: \code{LEFT}, \code{CENTER} and \code{RIGHT}}

\item{x}{any R object}
}
\value{
object of class BasicChartAxis
}
\description{
Specification for googlesheets Basic Chart
}
