% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{BasicChartDomain}
\alias{BasicChartDomain}
\alias{is.BasicChartDomain}
\title{BasicChartDomain}
\usage{
BasicChartDomain(gridRange, reversed = FALSE)

is.BasicChartDomain(x)
}
\arguments{
\item{gridRange}{specification for the cells. Only accepts one-row or
one-column cell ranges}

\item{reversed}{boolean indicating if values should be reversed}

\item{x}{any R object}
}
\description{
object of class BasicChartDomain. Specifies the context
of scopes (eg. date for numbers)
}
