% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BasicCharts.R
\name{BasicChartSeries}
\alias{BasicChartSeries}
\alias{is.BasicChartSeries}
\title{BasicChartSeries}
\usage{
BasicChartSeries(
  chartData,
  targetAxis = c("BOTTOM_AXIS", "LEFT_AXIS", "RIGHT_AXIS"),
  dataLabel = NULL,
  type = NULL,
  lineStyle = NULL,
  colorStyle = NULL,
  pointStyle = NULL,
  styleOverrides = NULL
)

is.BasicChartSeries(x)
}
\arguments{
\item{chartData}{object of class \link{ChartData} specifying the data scope.}

\item{targetAxis}{specifies on which axis the values will be visualised}

\item{dataLabel}{object of class \link{DataLabel} specifying labels for the series}

\item{type}{specifies the type of the series (relevant only if chart type is COMBO).
Possible values are \code{"LINE"}, \code{"AREA"} and \code{"COLUMN"}}

\item{lineStyle}{object of class \link{LineStyle}, which specifies style of the line.
Relevant only if chart type is one of: AREA, LINE, SCATTER or COMBO with correct
series \code{type}}

\item{colorStyle}{object of class \link{ColorStyle}, specifying color for
elements associated with the series.}

\item{pointStyle}{object of class \link{PointStyle}, specifying style of the data
points associated with the series. Valid only if chart type is one of: AREA,
LINE, SCATTER or COMBO with correct series \code{type}}

\item{styleOverrides}{currently unimplemented. Present for correct data
reading}

\item{x}{any R object}
}
\description{
object of class BasicChartDomain. Specifies the data
that will be visualized in the chart
}
