% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_cells.R
\name{CellRequests}
\alias{CellRequests}
\alias{UpdateCellsRequest}
\alias{AppendCellsRequest}
\title{Requests to update or append cells in the grid}
\usage{
UpdateCellsRequest(
  rows,
  fields = c("userEnteredValue", "userEnteredFormat", "note", "textFormatRuns",
    "dataValidation", "pivotTable", "dataSourceTable", "dataSourceFormula"),
  start = NULL,
  range = NULL
)

AppendCellsRequest(
  sheetId,
  rows,
  fields = c("userEnteredValue", "userEnteredFormat", "note", "textFormatRuns",
    "dataValidation", "pivotTable", "dataSourceTable", "dataSourceFormula")
)
}
\arguments{
\item{rows}{object of class \link{RowData} or list of multiple such objects}

\item{fields}{name of \link{CellData} fields to update}

\item{start}{object of class \link{GridCoordinate}. The coordinate to start writing
data at. Any number of rows and columns (including a different number of
columns per row) may be written. Can't be specified alongside \code{range.}}

\item{range}{object of class \link{GridRange} or list of such objects.
The range to write data to. If the data in rows does not cover the entire
requested range, the fields matching those set in fields will be cleared.
Can't be specified alongside \code{start}.}

\item{sheetId}{The sheet ID to append the data to.}
}
\value{
deepgsheets4Req object
}
\description{
Create \code{deepgsheets4Req} objects that allow for update or append
rows of cell data into sheet. Send created requests with \code{\link[=send_batchUpdate_req]{send_batchUpdate_req()}}
}
\section{UpdateCells}{

Update cells beginning at \code{start} (upper-left coordinate) or in \code{range}.
Only properties specified in \code{fields} will be updated. If a property
is specified in \code{fields} but is missing from \code{CellData}, the property
will be erased
}

\section{AppendCells}{

Adds new cells after the last row with data in a sheet, inserting new rows
into the sheet if necessary. Properties of \code{CellData} specified in
\code{fields} will be inserted.
}

\seealso{
Other deepgsheets4Req constructors: 
\code{\link{ConditionalFormatRequests}},
\code{\link{MergeRequests}},
\code{\link{PasteRequests}},
\code{\link{SheetRequests}}
}
\concept{deepgsheets4Req constructors}
