% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{ChartAxisViewWindowOptions}
\alias{ChartAxisViewWindowOptions}
\alias{is.ChartAxisViewWindowOptions}
\title{Options of view window in Chart Axis}
\usage{
ChartAxisViewWindowOptions(
  viewWindowMode = c("PRETTY", "EXPLICIT"),
  viewWindowMin = NULL,
  viewWindowMax = NULL
)

is.ChartAxisViewWindowOptions(x)
}
\arguments{
\item{viewWindowMode}{mode for view. \code{"PRETTY"} chooses min and max values
automatically}

\item{viewWindowMin, viewWindowMax}{minimum and maximum values shown on axis.
Used only when \code{viewWindowMode = "EXPLICIT"}}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{ChartAxisViewWindowOptions}
}
\description{
Object with specification on how the view on the chart should
be placed
}
\seealso{
Other Chart objects constructors: 
\code{\link{ChartData}()},
\code{\link{ChartSpec}()},
\code{\link{DataLabel}()},
\code{\link{DataSourceChartProperties}()},
\code{\link{EmbeddedChart}()}
}
\concept{Chart objects constructors}
