% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Charts.R
\name{ChartSpec}
\alias{ChartSpec}
\alias{is.ChartSpec}
\alias{is.TypeChartSpec}
\alias{is.EmbeddedObjectPosition}
\title{ChartSpec}
\usage{
ChartSpec(
  chart,
  title = NULL,
  titlePosition = c("CENTER", "LEFT", "RIGHT"),
  subtitle = NULL,
  subtitlePosition = c("CENTER", "LEFT", "RIGHT"),
  fontName = NULL,
  maximized = NULL
)

is.ChartSpec(x)

is.TypeChartSpec(x)

is.EmbeddedObjectPosition(x)
}
\arguments{
\item{chart}{Specification of the chart to render. Currently supports only
\code{BasicChartSpec} objects}

\item{title}{Title of the chart. Optional.}

\item{titlePosition}{Horizntal aligmnent of the title. Defaults to \code{CENTER}}

\item{subtitle}{Subtitle of the chart. Optional.}

\item{subtitlePosition}{Horizonstal aligment of the subtitle. Defaults to
\code{CENTER}}

\item{fontName}{name of the font to use. Optional.}

\item{maximized}{boolean. If \code{TRUE}, the chart will take maximum amount of
available space with minimal padding. Optional.}

\item{x}{any R object}
}
\description{
Specification for googlesheets chart
}
