% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConditionalFormatRequests.R
\name{ConditionalFormatRequests}
\alias{ConditionalFormatRequests}
\alias{AddConditionalFormatRule}
\alias{UpdateConditionalFormatRule}
\alias{DeleteConditionalFormatRule}
\title{Requests to add, modify and delete conditional formatting}
\usage{
AddConditionalFormatRule(index, rule)

UpdateConditionalFormatRule(
  index,
  rule = NULL,
  newIndex = NULL,
  sheetId = NULL
)

DeleteConditionalFormatRule(index, sheetId)
}
\arguments{
\item{index}{zero-based index identifying the rule}

\item{rule}{object of class \link{ConditionalFormatRule} describing the rule
for addition or update}

\item{newIndex}{zero-based index to which the rule should be moved}

\item{sheetId}{ID of the sheet the rule is located in}
}
\value{
deepgsheets4Req object
}
\description{
Create \code{deepgsheets4Req} objects that allow for addition, modification
and deletion of conditional formatting from given sheet. Send created
requests with \code{\link[=send_batchUpdate_req]{send_batchUpdate_req()}}
}
\section{Add}{

New rule is added at a given index. All subsequent rules indices are incremented.
}

\section{Update}{

Rule at a given index can be either replaced, providing the \code{rule} argument
or moved, providing the \code{newIndex} and \code{sheetId} arguments.
}

\section{Delete}{

Conditional format rule at given \code{index} from specified \code{sheetId} is
deleted, and all subsequent rules' indices are decremented.
}

\seealso{
Other deepgsheets4Req constructors: 
\code{\link{CellRequests}},
\code{\link{DimensionGroupRequests}},
\code{\link{DimensionRequests}},
\code{\link{MergeRequests}},
\code{\link{PasteRequests}},
\code{\link{SheetRequests}},
\code{\link{UpdateBordersRequest}()}
}
\concept{deepgsheets4Req constructors}
