% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{DataLabel}
\alias{DataLabel}
\alias{is.DataLabel}
\title{DataLabel}
\usage{
DataLabel(
  textFormat = NULL,
  placement = c("DATA_LABEL_PLACEMENT_UNSPECIFIED", "CENTER", "LEFT", "RIGHT", "ABOVE",
    "BELOW", "INSIDE_END", "INSIDE_BASE", "OUTSIDE_END"),
  type = c("DATA_LABEL_TYPE_UNSPECIFIED", "DATA", "CUSTOM", "NONE"),
  customLabelData = NULL
)

is.DataLabel(x)
}
\arguments{
\item{textFormat}{object of class \link{TextFormat} specifying formats of data
labels}

\item{placement}{placement of the data label relative to the series}

\item{type}{Type of the data label}

\item{customLabelData}{object of class \link{CellRange} specifying the data source
for data labels with \code{type = "CUSTOM"}}

\item{x}{any R object}
}
\value{
dgs4Obj of class \code{DataLabel}
}
\description{
Settings for chart data labels - annotations that appear on
the chart next to a series
}
\details{
\itemize{
\item \strong{Type} options:
\itemize{
\item DATA_LABEL_TYPE_UNSPECIFIED:  The data label type is not specified and
will be interpreted depending on the context of the data label within the chart.
\item DATA: The data label is displayed using values from the series data.
\item CUSTOM: The data label is displayed using values from a custom data source
indicated by \code{customLabelData}
\item NONE: The data label is not displayed
}
\item \strong{Placement} options:
\itemize{
\item DATA_LABEL_PLACEMENT_UNSPECIFIED: The positioning is determined
automatically by the renderer.
\item CENTER: Center within a bar or column, both horizontally and vertically.
\item LEFT: To the left of a data point.
\item RIGHT: To the right of a data point.
\item ABOVE: Above a data point.
\item BELOW: Below a data point.
\item INSIDE_END: Inside a bar or column at the end (top if positive, bottom if negative).
\item INSIDE_BASE: Inside a bar or column at the base.
\item OUTSIDE_END: Outside a bar or column at the end.
}
}
}
\seealso{
Other Chart objects constructors: 
\code{\link{ChartAxisViewWindowOptions}()},
\code{\link{ChartData}()},
\code{\link{ChartGroupRule}()},
\code{\link{ChartHistogramRule}()},
\code{\link{ChartSpec}()},
\code{\link{DataSourceChartProperties}()},
\code{\link{EmbeddedChart}()}
}
\concept{Chart objects constructors}
