% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dimension.R
\name{DimensionRange}
\alias{DimensionRange}
\alias{is.DimensionRange}
\title{Dimension Range}
\usage{
DimensionRange(
  sheetId,
  dimension = c("ROWS", "COLUMNS"),
  startIndex = NULL,
  endIndex = NULL
)

is.DimensionRange(x)
}
\arguments{
\item{sheetId}{ID of the sheet this range is on}

\item{dimension}{type of the dimension}

\item{startIndex, endIndex}{zero-based indices bounding the range. \code{startIndex}
is inclusive, \code{endIndex} is exclusive. Missing indices indicate the range is
unbounded on that side.}

\item{x}{any R object}
}
\description{
Object representing range alongside one of the dimensions.
}
