% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_dimensions.R
\name{DimensionRequests}
\alias{DimensionRequests}
\alias{InsertDimensionRequest}
\alias{AppendDimensionRequest}
\alias{UpdateDimensionPropertiesRequest}
\alias{AutoResizeDimensionsRequest}
\alias{DeleteDimensionRequest}
\alias{MoveDimensionRequest}
\title{Requests related to dimensions (rows/columns) in a sheet}
\usage{
InsertDimensionRequest(range, inheritFromBefore = NULL)

AppendDimensionRequest(sheetId, length, dimension = c("ROWS", "DIMENSIONS"))

UpdateDimensionPropertiesRequest(
  properties,
  fields = c("pixelSize", "developerMetadata", "hiddenByUser"),
  range = NULL,
  dataSourceSheetRange = NULL
)

AutoResizeDimensionsRequest(
  dimensions = NULL,
  dataSourceSheetDimensions = NULL
)

MoveDimensionRequest(source, destinationIndex)

DeleteDimensionRequest(range)
}
\arguments{
\item{range}{object of class \link{DimensionRange} declaring range of dimension
in \code{GRID} sheet}

\item{inheritFromBefore}{boolean indicating if the dimension properties
should be inherited from before the \code{range} (if TRUE) or from after the range
(FALSE)}

\item{sheetId}{ID of the sheet to append rows/columns}

\item{length}{number of rows/columns to append}

\item{dimension}{either \code{ROWS} or \code{COLUMNS} to check}

\item{properties}{object of class \link{DimensionProperties} providing data to
update}

\item{fields}{which fields of current properties need to be updated}

\item{dataSourceSheetRange}{object of class \link{DataSourceDimensionRange} specifying
the range of columns in \code{DATA_SOURCE} sheet}

\item{dimensions, DataSourceSheetDimensionRange}{objects of class, sequentially:
\link{DimensionRange} or \link{DataSourceDimensionRange}, declaring the range in \code{GRID}
or \code{DATA_SOURCE} sheet to resize automatically. Exactly one of these arguments
needs to be provided.}

\item{source}{object of class \link{DimensionRange} declaring the source range
for move}

\item{destinationIndex}{index of the first column or row to where the dimensions
should be moved. Provide the index as in \strong{before} the move.}
}
\value{
deepgsheets4Req object
}
\description{
Create \code{deepgsheets4Req} objects that allows inserting, appending, deletion
or updating a dimension in a sheet. Send created requests with
\code{\link[=send_batchUpdate_req]{send_batchUpdate_req()}}
}
\section{Insert}{

Inserts rows or columns in a sheet at a particular range Both \code{startIndex}
and \code{endIndex} must be bounded. Existing cells will be shifted to make place
for inserted dimension.
}

\section{Append}{

Appends given number of rows or columns to the end of a sheet.
}

\section{UpdateProperties}{

Send a request to update properties of dimension (column/row) in
a given range. Range can be specified with either \code{range} (for \code{GRID} sheets)
or \code{dataSourceSheetRange} (for \code{DATA_SOURCE} sheets) arguments
}

\section{AutoResize}{

Automatically resizes dimensions based on the contents of the cells in that
dimension.
}

\section{Move}{

Moves one or more rows or columns declared by \code{source} into destination
marked by \code{destinationIndex} (index of first row or column)
}

\section{Delete}{

Deletes the dimensions from the sheet in the given \code{range}. Other cells
will be shifted in their place.
}

\seealso{
Other deepgsheets4Req constructors: 
\code{\link{CellRequests}},
\code{\link{ConditionalFormatRequests}},
\code{\link{MergeRequests}},
\code{\link{PasteRequests}},
\code{\link{SheetRequests}}
}
\concept{deepgsheets4Req constructors}
