% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\name{EmbeddedChart}
\alias{EmbeddedChart}
\alias{is.EmbeddedChart}
\alias{gen_EmbeddedChart}
\title{EmbeddedChart}
\usage{
EmbeddedChart(spec, position, borderColor = NULL, chartId = NULL)

is.EmbeddedChart(x)

gen_EmbeddedChart(obj, sheetId = NULL)
}
\arguments{
\item{spec}{object of \link{ChartSpec} class, declaring chart specification}

\item{position}{object of \link{EmbeddedObjectPosition} class, declaring the
position of chart}

\item{borderColor}{object of \link{ColorStyle} class, declaring the color of
borders of the chart}

\item{chartId}{ID of the chart unique in a spreadsheet. If not set during
write, the one is chosen for you.}

\item{x}{any R object}

\item{obj}{list produced by \code{deepgs_listinize()}}

\item{sheetId}{optional sheetId}
}
\description{
Specification of a chart embedded in a sheet
}
