% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmbeddedObject.R
\name{EmbeddedObjectPosition}
\alias{EmbeddedObjectPosition}
\alias{is.EmbeddedObjectPosition}
\alias{gen_EmbeddedObjectPosition}
\title{EmbeddedObjectPosition}
\usage{
EmbeddedObjectPosition(overlayPosition = NULL, sheetId = NULL, newSheet = TRUE)

is.EmbeddedObjectPosition(x)

gen_EmbeddedObjectPosition(obj, sheetProperties = NULL)
}
\arguments{
\item{overlayPosition}{object of class \link{OverlayPosition} declaring position
of EmbeddedObject in existing sheet}

\item{sheetId}{the ID of sheet this object is in if EmbeddedObject is in
its own sheet during read or update. When creating new object, it specifies
ID of new sheet that will be created.}

\item{newSheet}{TRUE if the object is to be put on a new sheet. Its ID is
chosen for you. Used only when creating new object.}

\item{x}{any R object}

\item{obj}{list produced by \code{deepgs_listinize()}}

\item{sheetProperties}{optional \code{SheetProperties} object to get additional
data during read from API}
}
\value{
EmbeddedObjectPosition
}
\description{
Specification of position for embedded objects (eg.
charts and slicers). Out of three arguments, only one can be provided. They
are interpreted in order.
}
