% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generators.R
\name{gen_dgs4Resp}
\alias{gen_dgs4Resp}
\alias{@eval}
\alias{pkg_env$resp_generators}
\alias{gen_AddSheet}
\alias{gen_DuplicateSheet}
\alias{gen_AddChart}
\title{Generate dgs4Resp}
\usage{
gen_dgs4Resp(obj, type, sheetId = NULL)

gen_AddSheet(obj)

gen_DuplicateSheet(obj)

gen_AddChart(obj, sheetId = NULL)
}
\arguments{
\item{obj}{list generated by \code{\link[=dgs4_listinize]{dgs4_listinize()}} or list received from
googlesheets API}

\item{type}{name of the response type}

\item{sheetId}{optional sheetId to paste into the constructor. Necessary for
some objects returned from googlesheets API}
}
\value{
dgs4Resp object
}
\description{
Responses of \link{request_ss_batchUpdate} returned from Sheets API are coerced
into \code{dgs4Resp} objects, allowing for dropping unnecessary nesting and
transforming returned objects into \code{dgs4Obj} classes.
}
