% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridCoordinate}
\alias{GridCoordinate}
\alias{is.GridCoordinate}
\alias{gen_GridCoordinate}
\title{GridCoordinate}
\usage{
GridCoordinate(sheetId, rowIndex, columnIndex)

is.GridCoordinate(x)

gen_GridCoordinate(obj, sheetId = NULL)
}
\arguments{
\item{sheetId}{optional sheetId}

\item{rowIndex, columnIndex}{position of the cell in a grid}

\item{x}{any R object}

\item{obj}{list produced by \code{deepgs_listinize()}}
}
\description{
Specification of cell in a grid
}
\section{Warning}{

Objects representing \code{GridCoordinate} received from GoogleSheets API v4
don't include \code{sheetId}. It is filled during response processing with \code{sheetId}
of the sheet in which the object was located in.
}

