% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridCoordinate}
\alias{GridCoordinate}
\alias{is.GridCoordinate}
\alias{print.GridCoordinate}
\title{GridCoordinate}
\usage{
GridCoordinate(sheetId, rowIndex, columnIndex)

is.GridCoordinate(x)

\method{print}{GridCoordinate}(gc, A1 = getOption("deepgs4.A1", FALSE), ...)
}
\arguments{
\item{sheetId}{Integer - sheet ID}

\item{rowIndex, columnIndex}{position of the cell in a grid}

\item{x}{any R object}

\item{gc}{GridCoordinate object}

\item{A1}{if \code{TRUE} then include \emph{A1} notation. Can be set globally with
\code{options("deepgs4.A1" = TRUE)}}

\item{...}{optional arguments to \code{print} methods.}
}
\description{
Specification of cell in a grid
}
\section{Warning}{

Objects representing \code{GridCoordinate} received from GoogleSheets API v4
don't include \code{sheetId}. It is filled during response processing with \code{sheetId}
of the sheet in which the object was located in.
}

