% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridProperties}
\alias{GridProperties}
\alias{is.GridProperties}
\title{GridProperties}
\usage{
GridProperties(
  rowCount,
  columnCount,
  frozenRowCount = NULL,
  frozenColumnCount = NULL,
  hideGridlines = NULL,
  rowGroupControlAfter = NULL,
  columnGroupControlAfter = NULL
)

is.GridProperties(x)
}
\arguments{
\item{rowCount, columnCount}{number of rows and columns in the grid}

\item{frozenRowCount, frozenColumnCount}{number of frozen rows and columns in the grid}

\item{rowGroupControlAfter, columnGroupControlAfter}{boolean indicating if
row/column toggle is shown after the group}

\item{x}{any R object}

\item{hideGridLines}{boolean indicating if gridlines should be shown}
}
\description{
Properties of a chart grid
}
