% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridProperties}
\alias{GridProperties}
\alias{gen_GridProperties}
\title{GridProperties}
\usage{
GridProperties(
  rowCount,
  columnCount,
  frozenRowCount = NULL,
  frozenColumnCount = NULL,
  hideGridLines = NULL,
  rowGroupControlAfter = NULL,
  columnGroupControlAfter = NULL
)

gen_GridProperties(obj)
}
\arguments{
\item{rowCount, columnCount}{number of rows and columns in the grid}

\item{frozenRowCount, frozenColumnCount}{number of frozen rows and columns in the grid}

\item{hideGridLines}{boolean indicating if gridlines should be shown}

\item{rowGroupControlAfter, columnGroupControlAfter}{boolean indicating if
row/column toggle is shown after the group}

\item{obj}{list produced by \code{deepgs_listinize()}}
}
\description{
Properties of a chart grid
}
