% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridRange}
\alias{GridRange}
\alias{is.GridRange}
\title{GridRange}
\usage{
GridRange(
  sheetId,
  startRowIndex,
  endRowIndex,
  startColumnIndex,
  endColumnIndex
)

is.GridRange(x)
}
\arguments{
\item{sheetId}{Integer - sheet ID}

\item{startRowIndex}{Integer. Starts from 0, inclusive}

\item{endRowIndex}{Integer. Starts from 0, exclusive}

\item{startColumnIndex}{Integer. Starts from 0, inclusive}

\item{endColumnIndex}{Integer. Starts from 0, exclusive}

\item{x}{any R object}
}
\description{
Specification of grid range in spreadsheet
}
\details{
Object representing \code{GridRange} received from GoogleSheets API v4
don't include \code{sheetId}. It is filled during response processing with \code{Id}
of the sheet which the higher-level object was located in.
}
