% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GridRange.R
\name{GridRange}
\alias{GridRange}
\alias{is.GridRange}
\title{GridRange}
\usage{
GridRange(
  sheetId,
  startRowIndex,
  endRowIndex,
  startColumnIndex,
  endColumnIndex
)

is.GridRange(x)
}
\arguments{
\item{sheetId}{Integer. Starts from 0. Spreadsheet Id}

\item{startRowIndex}{Integer. Starts from 0, inclusive}

\item{endRowIndex}{Integer. Starts from 0, exclusive}

\item{startColumnIndex}{Integer. Starts from 0, inclusive}

\item{endColumnIndex}{Integer. Starts from 0, exclusive}

\item{x}{any R object}
}
\description{
Specification of grid range in spreadsheet
}
