% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Grid.R
\name{GridRange}
\alias{GridRange}
\alias{is.GridRange}
\alias{gen_GridRange}
\title{GridRange}
\usage{
GridRange(
  sheetId,
  startRowIndex,
  endRowIndex,
  startColumnIndex,
  endColumnIndex
)

is.GridRange(x)

gen_GridRange(obj, sheetId = NULL)
}
\arguments{
\item{sheetId}{optional sheetId}

\item{startRowIndex}{Integer. Starts from 0, inclusive}

\item{endRowIndex}{Integer. Starts from 0, exclusive}

\item{startColumnIndex}{Integer. Starts from 0, inclusive}

\item{endColumnIndex}{Integer. Starts from 0, exclusive}

\item{x}{any R object}

\item{obj}{list produced by \code{deepgs_listinize()}}
}
\description{
Specification of grid range in spreadsheet
}
\section{Warning}{

Objects representing \code{GridRange} received from GoogleSheets API v4
don't include \code{sheetId}. It is filled during response processing with \code{sheetId}
of the sheet in which the object was located in.
}

