% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spreadsheet.R
\name{IterativeCalculationSettings}
\alias{IterativeCalculationSettings}
\alias{gen_IterativeCalculationSettings}
\title{Settings of iterative calculations}
\usage{
IterativeCalculationSettings(maxIterations, convergenceThreshold = NULL)

gen_IterativeCalculationSettings(obj)
}
\arguments{
\item{maxIterations}{When iterative calculation is enabled, the maximum
number of calculation rounds to perform.}

\item{convergenceThreshold}{When iterative calculation is enabled and
successive results differ by less than this threshold value, the calculation
rounds stop.}

\item{obj}{list generated by \code{\link[=deepgs_listinize]{deepgs_listinize()}}}
}
\description{
Settings to control how circular dependencies are resolved with
iterative calculation.
}
