% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_merge.R
\name{MergeCellsRequest}
\alias{MergeCellsRequest}
\title{Merge cells request}
\usage{
MergeCellsRequest(
  gridRange,
  mergeType = c("MERGE_COLUMNS", "MERGE_ROWS", "MERGE_ALL")
)
}
\arguments{
\item{gridRange}{object of class \link{GridRange} declaring the range in which
cells need to be merged}

\item{mergeType}{one of MERGE_COLUMNS, MERGE_ROWS or MERGE_ALL}
}
\value{
deepgsheets4Req
}
\description{
Single MergeCells request
}
