% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_requests.R
\name{MergeCellsRequest}
\alias{MergeCellsRequest}
\title{MergeCells}
\usage{
MergeCellsRequest(
  gridRange,
  mergeType = c("MERGE_COLUMNS", "MERGE_ROWS", "MERGE_ALL")
)
}
\arguments{
\item{mergeType}{one of MERGE_COLUMNS, MERGE_ROWS or MERGE_ALL}

\item{GridRange}{\code{GridRange} object}
}
\value{
\code{deepgsheets4Req} object
}
\description{
Single MergeCells request
}
