% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cells.R
\name{Padding}
\alias{Padding}
\alias{gen_Padding}
\title{Cell padding}
\usage{
Padding(top, bottom, left, right)

gen_Padding(obj)
}
\arguments{
\item{top, bottom, left, right}{Padding in pixels for each side. During update,
all fields need to be set}

\item{obj}{list produced by \code{deepgs_listinize()}}
}
\description{
Cell padding
}
