% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendRequest.R
\name{send_batchUpdate_req}
\alias{send_batchUpdate_req}
\title{Send \emph{batchUpdate} request to googlesheets API}
\usage{
send_batchUpdate_req(spreadsheetId, ..., .dots = list())
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet to apply the changes}

\item{...}{objects of class \link{deepgsheets4Req}}

\item{.dots}{the same as \code{...}, but applied as a list}
}
\description{
This function sends \emph{batchUpdate} request to googlesheets API
based on multiple \link{deepsheets4Req} objects
}
