% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sheet.R
\name{Sheet}
\alias{Sheet}
\alias{is.Sheet}
\title{Googlesheets Sheet specification}
\usage{
Sheet(
  properties = NULL,
  data = NULL,
  merges = NULL,
  conditionalFormats = NULL,
  filterViews = NULL,
  protectedRanges = NULL,
  basicFilter = NULL,
  charts = NULL,
  bandedRanges = NULL,
  developerMetadata = NULL,
  rowGroups = NULL,
  columnGroups = NULL,
  slicers = NULL
)

is.Sheet(x)
}
\arguments{
\item{x}{any R object}
}
\description{
Specification of singular Sheet of the google spreadsheeet
}
