% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_sheet.R
\name{SheetRequests}
\alias{SheetRequests}
\alias{AddSheetRequest}
\alias{UpdateSheetPropertiesRequest}
\alias{DeleteSheet}
\alias{DeleteSheetRequest}
\title{Requests to add, modify and delete Sheet}
\usage{
AddSheetRequest(properties)

UpdateSheetPropertiesRequest(
  properties,
  fields = c("title", "index", "gridProperties", "hidden", "tabColorStyle",
    "rightToLeft")
)

DeleteSheetRequest(sheetId)
}
\arguments{
\item{properties}{object of class \link{SheetProperties}}

\item{fields}{which of the properties specified within \code{properties} should
be updated}

\item{sheetId}{integer representing the sheet to delete}
}
\value{
deepgsheets4Req object
}
\description{
Create \code{deepgsheets4Req} objects that allow for addition, modification
and deletion of singular sheet. Send created requests with \code{\link[=send_batchUpdate_req]{send_batchUpdate_req()}}
}
\section{Add}{

Adds a new sheet. When a sheet is added at a given index,
all subsequent sheets' indexes are incremented. All properties are optional.
The sheetId field is optional; if one is not set, an ID will be randomly generated.
}

\section{Update}{

\code{sheetId} field of the \code{properties} will be used to identify which sheet
properties need to be updated. Only properties specified in \code{fields} will
be updated.
}

\section{Delete}{

Sheet of the given \code{sheetId} will be deleted. If the sheet is of \code{DATA_SOURCE}
type, the associated DataSource is also deleted.
}

\seealso{
Other deepgsheets4Req constructors: 
\code{\link{CellRequests}},
\code{\link{ConditionalFormatRequests}},
\code{\link{DimensionRequests}},
\code{\link{MergeRequests}},
\code{\link{PasteRequests}}
}
\concept{deepgsheets4Req constructors}
