% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spreadsheet.R
\name{Spreadsheet}
\alias{Spreadsheet}
\title{GoogleSheet Soreadsheet}
\usage{
Spreadsheet(
  spreadsheetId,
  properties = NULL,
  sheets = NULL,
  namedRanges = NULL,
  developerMetadata = NULL,
  dataSources = NULL,
  spreadsheetUrl = NULL,
  dataSourceSchedules = NULL
)
}
\arguments{
\item{spreadsheetId}{ID of the spreadsheet}

\item{properties}{object of class \link{SpreadsheetProperties}}

\item{sheets}{object of class \link{SheetProperties} or list of such objects}

\item{namedRanges}{object of class \link{NamedRange} or list of such objects}

\item{developerMetadata}{object of class \link{DeveloperMetadata} or list of such
objects.}

\item{dataSources}{object of class \link{DataSource} or list of such objects.
Represents external data sources for the spreadsheet}

\item{spreadsheetUrl}{\strong{READ ONLY} spreadsheet URL}

\item{dataSourceSchedules}{\strong{READ ONLY} object of class \link{DataSourceRefreshSchedule}
or list of such objects. Represents schedules for refreshment of data sources.}
}
\description{
GoogleSheet Soreadsheet
}
\details{
Object representing complete spreadsheet
}
