% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spreadsheet.R
\name{SpreadsheetTheme}
\alias{SpreadsheetTheme}
\alias{gen_SpreadsheetTheme}
\title{Spreadsheet Theme}
\usage{
SpreadsheetTheme(
  primaryFontFamily = NULL,
  TEXT = NULL,
  BACKGROUND = NULL,
  ACCENT1 = NULL,
  ACCENT2 = NULL,
  ACCENT3 = NULL,
  ACCENT4 = NULL,
  ACCENT5 = NULL,
  ACCENT6 = NULL,
  LINK = NULL
)

gen_SpreadsheetTheme(obj)
}
\arguments{
\item{primaryFontFamily}{Name of the primary font family.}

\item{TEXT, BACKGROUND, ACCENT1, ACCENT2, ACCENT3, ACCENT4, ACCENT5, ACCENT6, LINK}{Objects of \code{\link[=ColorStyle]{ColorStyle()}} class with explicit \emph{red}, \emph{green}, \emph{blue}
specification for each of the \emph{theme colors}. During update all themes need
to be specified.}

\item{obj}{list generated by \code{\link[=deepgs_listinize]{deepgs_listinize()}}}
}
\description{
Specification of spreadsheet theme
}
