% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Varia.R
\name{TextFormatRun}
\alias{TextFormatRun}
\title{TextFormat for part of the cell value}
\usage{
TextFormatRun(startIndex = NULL, textFormat)
}
\arguments{
\item{startIndex}{index (zero-based) of character on which this format run.
Can be kept \code{NULL} for beginning of the value.}

\item{textFormat}{object of class \link{TextFormat} with format specification to
apply}
}
\description{
Object holding specification for usage with \code{\link[=CellData]{CellData()}} to
keep information about only part of a cell. The format of this run continues
until the start index of the next run.
}
