% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Chart.R
\docType{data}
\name{TypeChartSpecs}
\alias{TypeChartSpecs}
\title{Valid Chart Specs}
\format{
An object of class \code{character} of length 1.
}
\usage{
TypeChartSpecs
}
\description{
List of valid types of charts that can be provided to \code{chart}
argument of \code{\link[=ChartSpec]{ChartSpec()}} and, consequently created by sending \code{\link[=AddChartRequest]{AddChartRequest()}}
}
\keyword{datasets}
