% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_merge.R
\name{UnmergeCellsRequest}
\alias{UnmergeCellsRequest}
\title{Unmerge cells request}
\usage{
UnmergeCellsRequest(gridRange)
}
\arguments{
\item{gridRange}{object of class \link{GridRange} declaring the cells to be unmerged}
}
\value{
deepgsheets4Req
}
\description{
Request unmerging all merged cells in given range
}
