% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_chart.R
\name{UpdateChartSpecRequest}
\alias{UpdateChartSpecRequest}
\title{Update Chart Specification}
\usage{
UpdateChartSpecRequest(chartId, spec)
}
\arguments{
\item{chartId}{id of the chart to update}

\item{spec}{object of class \link{ChartSpec}, containing new chart specifications}
}
\value{
deepgsheets4Req
}
\description{
Constructs request for updating chart specification
}
\seealso{
Other deepgsheets4Req: 
\code{\link{AddChartRequest}()},
\code{\link{MergeCellsRequest}()},
\code{\link{UnmergeCellsRequest}()},
\code{\link{is.deepgsheets4Req}()}
}
\concept{deepgsheets4Req}
