% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_fields_validity.R
\name{valid_update_fields}
\alias{valid_update_fields}
\title{Check validity of specified fields}
\usage{
valid_update_fields(updateRequest = c("UpdateCells", "UpdateSheetProperties"))
}
\arguments{
\item{updateRequest}{which update request to check}

\item{fields}{character vector in}
}
\value{
character vector
}
\description{
Many requests send to Sheets API take
a \code{fields} argument in form of \href{https://developers.google.com/protocol-buffers/docs/reference/google.protobuf#google.protobuf.FieldMask}{FieldMask}{target="_blank"}

\emph{FieldMask} format is helpful for specifying accurately which fields should
be updated by batchUpdate request or retrieved with \emph{get} request

This function works as a helper for retrieval of valid update fields per
request.
}
